% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf.R
\name{ACF}
\alias{ACF}
\title{Auto-Covariance and Correlation Functions}
\usage{
ACF(x, lagmax = 0, cor = TRUE, demean = TRUE)
}
\arguments{
\item{x}{A \code{matrix} with dimensions \eqn{N \times S}{N x S} or N observations and S processes}

\item{lagmax}{A \code{integer}}

\item{cor}{A \code{bool} indicating whether the correlation 
(\code{TRUE}) or covariance (\code{FALSE}) should be computed.}

\item{demean}{A \code{bool} indicating whether the data should be detrended
(\code{TRUE}) or not (\code{FALSE})}
}
\value{
An \code{array} of dimensions \eqn{N \times S \times S}{N x S x S}.
}
\description{
The acf function computes the estimated
autocovariance or autocorrelation for both univariate and multivariate cases.
}

