% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{acf_sum}
\alias{acf_sum}
\title{Helper Function for ARMA to WV Approximation}
\usage{
acf_sum(ar, ma, last_tau, alpha = 0.99)
}
\arguments{
\item{ar}{A \code{vec} containing the coefficients of the AR process}

\item{ma}{A \code{vec} containing the coefficients of the MA process}

\item{last_tau}{An \code{int} the Jth scale of 2^(1:J)}

\item{alpha}{A \code{double} indicating the cutoff.}
}
\value{
A \code{vec} containing the wavelet variance of the ARMA process.
}
\description{
Indicates where the minimum ARMAacf value is and returns that as an index.
}
\seealso{
\code{\link{arma_to_wv_app}}
}
\keyword{internal}
