% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{addUnits}
\alias{addUnits}
\title{Add Units to Sensor}
\usage{
addUnits(units, sensor)
}
\arguments{
\item{units}{\code{NULL} or a two-element vector indicating the units of gyroscope and accelerometer sensors.}

\item{sensor}{A \code{character vector} including all sensors in imu objects.}
}
\description{
Add corresponding units for gyroscope and accelerometer sensors.
}
\details{
This function is used in \code{plot.wvar.imu} and \code{plot.auto.imu}.
}
\keyword{internal}
