% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{adv_gmwm_imu_ssm_cpp}
\alias{adv_gmwm_imu_ssm_cpp}
\title{User Specified Initial Values for GMWM Estimator}
\usage{
adv_gmwm_imu_ssm_cpp(theta, desc, model_type, V, wv_empir, tau, N)
}
\arguments{
\item{theta}{A \code{vector} with dimensions N x 1 that contains user-supplied initial values for parameters}

\item{desc}{A \code{vector<string>} indicating the models that should be considered.}

\item{V}{A \code{matrix} that represents the covariance matrix.}

\item{wv_empir}{A \code{vector} that contains the empirical wavelet variance}

\item{N}{A \code{integer} that indicates the length of the signal being studied.}
}
\value{
A \code{vec} that contains the parameter estimates from GMWM estimator.
}
\description{
This function uses the Generalized Method of Wavelet Moments to estimate the parameters of a time series model.
}
\details{
The function estimates a variety of time series models. If type = "ARMA" then the parameter vector (param) should
indicate the order of the AR process and of the MA process (i.e. param = c(AR,MA)). If type = "IMU" or "SSM", then
parameter vector should indicate the characters of the models that compose the latent or state-space model. The model
options are:
\itemize{
  \item{"AR1"}{a first order autoregressive process with parameters \eqn{(\phi,\sigma^2)}{phi, sigma^2}}
  \item{"DR"}{a drift with parameter \eqn{\omega}{omega}}
  \item{"QN"}{a quantization noise process with parameter \eqn{Q}}
  \item{"RW"}{a random walk process with parameter \eqn{\sigma^2}{sigma^2}}
  \item{"WN"}{a white noise process with parameter \eqn{\sigma^2}{sigma^2}}
}
If type = "ARMA", the function takes condition least squares as starting values; if type = "IMU" or type = "SSM" then
starting values pass through an initial bootstrap and pseudo-optimization before being passed to the GMWM optimization.
If robust = TRUE the function takes the robust estimate of the wavelet variance to be used in the GMWM estimation procedure.
}
\examples{
# Coming soon
}
\author{
JJB
}
\references{
Wavelet variance based estimation for composite stochastic processes, S. Guerrier and Robust Inference for Time Series Models: a Wavelet-Based Framework, S. Guerrier
}
\keyword{internal}

