% Generated by roxygen2 (4.0.2): do not edit by hand
\name{allan_variance}
\alias{allan_variance}
\title{Compute Allan Variance}
\usage{
allan_variance(x)
}
\arguments{
\item{x}{A \code{vector} with dimensions N x 1.}
}
\value{
av A \code{matrix} that contains the cluster size as the first column and the allan variance as the second.
}
\description{
Computation of allan variance
}
\examples{
white.noise = rnorm(500, 0, 1)
allan_variance(white.noise)
}
\author{
JJB
}

