% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ts.model.R
\name{AR1}
\alias{AR1}
\title{Create an Autoregressive 1 [AR(1)] Process}
\usage{
AR1(phi = 0.91, sigma = 1)
}
\arguments{
\item{phi}{A \code{double} value for the \eqn{\phi}{phi} of an AR1 process.}

\item{sigma}{A \code{double} value for the \eqn{\sigma}{sigma} process.}
}
\value{
An S3 object with called ts.model with the following structure:
\itemize{
 \item{desc}{"AR1"}
 \item{theta}{\eqn{phi}{\phi}, \eqn{sigma}{\sigma}}
}
}
\description{
Setups the necessary backend for the AR1 process.
}
\details{
A standard deviation is required since the model generation statements utilize
randomization functions expecting a standard deviation instead of a variance.
}
\examples{
AR1()
AR1(phi=.32, sigma=1.3)
}
\author{
JJB
}

