% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ar1_to_wv}
\alias{ar1_to_wv}
\title{AR1 process to WV}
\usage{
ar1_to_wv(phi, sig2, Tau)
}
\arguments{
\item{Tau}{A \code{vec} containing the scales e.g. 2^tau}

\item{omega}{A \code{double} corresponding to variance of drift}
}
\value{
A \code{vec} containing the wavelet variance of the drift.
}
\description{
This function compute the WV (haar) of an AR(1) process
}
\examples{
x.sim = gen_ar1( N = 10000, phi = 0.9, sig2 = 4 )
waveletVariance(x.sim)
wv.theo = ar1_to_wv(phi = 0.9, sig2 = 16, floor(log(length(x.sim),2)))
lines(wv.theo$scales,wv.theo$WV, col = "red")
}

