% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{arma11_to_wv}
\alias{arma11_to_wv}
\title{ARMA(1,1) to WV}
\usage{
arma11_to_wv(phi, theta, sig2, tau)
}
\arguments{
\item{phi}{A \code{double} corresponding to the autoregressive term.}

\item{theta}{A \code{double} corresponding to the moving average term.}

\item{sig2}{A \code{double} the variance of the process.}

\item{tau}{A \code{vec} containing the scales e.g. 2^tau}
}
\value{
A \code{vec} containing the wavelet variance of the ARMA(1,1) process.
}
\description{
This function computes the WV (haar) of an Autoregressive Order 1 - Moving Average Order 1 (ARMA(1,1)) process.
}
\details{
This function is significantly faster than its generalized counter part
\code{\link[gmwm]{arma_to_wv}}
}
\examples{
ntau = 7
tau = 2^(1:ntau)
wv.theo = arma11_to_wv(0.3, 0.1, 1, tau)
}

