% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{arma_to_wv}
\alias{arma_to_wv}
\title{ARMA process to WV}
\usage{
arma_to_wv(ar, ma, sigma2, tau)
}
\arguments{
\item{ar}{A \code{vec} containing the coefficients of the AR process}

\item{ma}{A \code{vec} containing the coefficients of the MA process}

\item{sigma2}{A \code{double} containing the residual variance}

\item{tau}{A \code{vec} containing the scales e.g. 2^tau}
}
\value{
A \code{vec} containing the wavelet variance of the ARMA process.
}
\description{
This function computes the (haar) WV of an ARMA process
}
\examples{

# Calculates the Haar WV for an ARMA(2,3).
arma_to_wv(c(.23,.43), c(.34,.41,.59), 3, 2^(1:9))
}
\seealso{
\code{\link{ARMAtoMA_cpp}},\code{\link{ARMAacf_cpp}}
}

