% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GMWM.R
\name{autoplot.comp}
\alias{autoplot.comp}
\title{Compare GMWM Model Fits on Same Graph}
\usage{
\method{autoplot}{comp}(object, ...)
}
\arguments{
\item{object}{A \code{data.frame} containing both sets of GMWM object data.}

\item{...}{other arguments passed to specific methods}
}
\value{
A ggplot2 panel containing one graphs with two GMWM models plotted against each other.
}
\description{
Creates a single graph that contains two GMWM models plotted against each other.
}
\examples{
\dontrun{# AR
set.seed(1335)
n = 200
x = gen.ts(AR1(phi = .1, sigma2 = 1) + AR1(phi = 0.95, sigma2 = .1), n)
GMWM1 = gmwm(AR1(), data = x)
GMWM2 = gmwm(2*AR1(), data = x)
compare.models(GMWM1, GMWM2, split = FALSE)}
}
\author{
JJB
}

