% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GMWM.R
\name{autoplot.compSplit}
\alias{autoplot.compSplit}
\title{Compare GMWM Model Fit on Split Graphs}
\usage{
\method{autoplot}{compSplit}(object, ...)
}
\arguments{
\item{object}{A \code{data.frame} containing both sets of GMWM object data.}

\item{...}{other arguments passed to specific methods}
}
\value{
A ggplot2 panel containing two graphs of the wavelet variance.
}
\description{
Creates GMWM model fits of two models on split graphs within the same panel.
}
\examples{
\dontrun{# AR
set.seed(1355)
n = 200
x = gen_ar1(n, phi=.1, sigma2 = 1) + gen_ar1(n,phi=0.95, sigma2 = .1)
decomp = modwt(x)
wv = wvar(decomp, robust = FALSE)
out = wvcov(decomp, wv, compute.v="diag")
GMWM1 = gmwm(AR1(), wvcov=out, signal=x, model.type="imu", B = 10000)
GMWM2 = gmwm(2*AR1(), wvcov=out, signal=x, model.type="imu", B = 10000)
compare.models(GMWM1, GMWM2, split = TRUE)}
}
\author{
JJB
}

