% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GMWM.R
\name{autoplot.gmwm}
\alias{autoplot.gmwm}
\title{Graph Solution of the Generalized Method of Wavelet Moments}
\usage{
\method{autoplot}{gmwm}(object, ...)
}
\arguments{
\item{object}{A \code{GMWM} object}

\item{...}{other arguments passed to specific methods}
}
\value{
A ggplot2 panel containing the graph of the empirical and theoretical wavelet variance under the constructed GMWM.
}
\description{
Creates a graph containing the empirical and theoretical wavelet variances constructed via GMWM.
}
\examples{
# AR
set.seed(1336)
n = 200
x = gen.ts(AR1(phi = .1, sigma2 = 1) + AR1(phi = 0.95, sigma2 = .1), n)
mod = gmwm(2*AR1(), data = x)
autoplot(mod)
}
\author{
JJB
}

