% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GMWM.R
\name{autoplot.gmwm2}
\alias{autoplot.gmwm2}
\title{Graph Solution of the Generalized Method of Wavelet Moments}
\usage{
\method{autoplot}{gmwm2}(object, ...)
}
\arguments{
\item{object}{A \code{GMWM2} object.}

\item{...}{other arguments passed to specific methods}
}
\value{
A ggplot2 panel containing the graph of the empirical and theoretical wavelet variance under the constructed GMWM for each latent process.
}
\description{
Creates a graph containing the empirical and theoretical wavelet variances constructed via GMWM for each latent process.
}
\author{
JJB
}

