% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GMWM.R
\name{autoplot.gmwmComp}
\alias{autoplot.gmwmComp}
\title{Compare GMWM Model Fits with \code{ggplot2}}
\usage{
\method{autoplot}{gmwmComp}(object, breaks, split = TRUE, CI = TRUE,
  background = "white", transparence = 0.1, line.color = NULL,
  CI.color = NULL, line.type = NULL, point.size = NULL,
  point.shape = NULL, title = "Haar Wavelet Variance Representation",
  title.size = 15, axis.label.size = 13, axis.tick.size = 11,
  axis.x.label = expression(paste("Scale ", tau)),
  axis.y.label = expression(paste("Wavelet Variance ", nu)),
  facet.label.size = 13, legend.label = NULL, legend.title = "",
  legend.key.size = 1.3, legend.title.size = 13, legend.text.size = 13,
  nrow = 1, ...)
}
\arguments{
\item{object}{A \code{data.frame} containing both sets of GMWM object data.}

\item{...}{other arguments passed to specific methods}
}
\value{
A ggplot2 panel containing one graph with several GMWM models plotted against each other.
}
\description{
Creates a single graph that contains several GMWM models plotted against each other.
}
\author{
JJB, Wenchao
}

