% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/imu_graphs.r
\name{autoplot.imu2}
\alias{autoplot.imu2}
\title{Plot in combined type: 2 graphs}
\usage{
\method{autoplot}{imu2}(object, CI = T, transparence = 0.1,
  line.type = "solid", graph.title = "Haar Wavelet Variance Representation",
  graph.title.size = 15, axis.label.size = 13, axis.tick.size = 11,
  title.x.axis = expression(paste("Scale ", tau)),
  title.y.axis = expression(paste("Wavelet Variance ", nu)),
  facet.title.size = 13, legend.title = "Axis", legend.key.size = 1.3,
  legend.title.size = 13, legend.text.size = 13, ...)
}
\arguments{
\item{object}{An \code{imu} object}

\item{CI}{A \code{boolean} that indicates whether the confidence interval should be plotted.}

\item{transparence}{A \code{double} that ranges from 0 to 1 that controls the transparency of the graph}

\item{line.type}{A \code{string} that indicates the type of line (e.g. solid, dotted, etc.)}

\item{graph.title}{A \code{string} that indicates the title of the graph}

\item{graph.title.size}{An \code{integer} that indicates the size of title.}

\item{axis.label.size}{An \code{integer} that indicates the size of label}

\item{axis.tick.size}{An \code{integer} that indicates the size of tick mark}

\item{title.x.axis}{A \code{string} that indicates the label on x axis}

\item{title.y.axis}{A \code{string} that indicates the label on y axis}

\item{facet.title.size}{An \code{integer} that indicates the size of facet label}

\item{legend.title}{A \code{string} that indicates the title of legend}

\item{legend.key.size}{A \code{double} that indicates the size of key (in centermeters) on legend}

\item{legend.title.size}{An \code{integer} that indicates the size of title on legend}

\item{legend.text.size}{An \code{integer} that indicates the size of key label on legend}

\item{...}{Additional options}
}
\value{
A panel containing the split graphs of an IMU sensor.
}
\description{
Plot each WV variance in a separate graph
}
\examples{
\dontrun{
data(imu)
df = imu2WV(imu)
plot(df, separate=FALSE)
}
}

