% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/imu_graphs.r
\name{autoplot.imu6}
\alias{autoplot.imu6}
\title{Plot in separate type: 6 graphs}
\usage{
\method{autoplot}{imu6}(object, transparence = 0.1, point.size = 0,
  color.line = "black", color.point = "#003C7D", color.CI = "#003C7D",
  line.type = "solid", ...)
}
\arguments{
\item{object}{An \code{imu} object}

\item{transparence}{A \code{double} that ranges from 0 to 1 that controls the transparency of the graph}

\item{point.size}{An \code{integer} that indicates the size of the points.}

\item{color.line}{A \code{string} that indicates the color of the line drawn.}

\item{color.point}{A \code{string} that is a hexadecimal color value.}

\item{color.CI}{A \code{string} that is a hexadecimal color value.}

\item{line.type}{A \code{string} that indicates the type of line (e.g. solid, dotted, etc.)}

\item{...}{Additional options}
}
\value{
A panel containing the split graphs of an IMU sensor.
}
\description{
Plot each WV variance in a separate graph
}
\examples{
\dontrun{
data(imu)
df = imu2WV(imu)
plot(df)
}
}

