% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wvar.R
\name{autoplot.wvar}
\alias{autoplot.wvar}
\title{Graph Wavelet Variances}
\usage{
\method{autoplot}{wvar}(object, ...)
}
\arguments{
\item{object}{A \code{wvar} object.}

\item{...}{other arguments passed to specific methods}
}
\value{
A ggplot2 graph containing the wavelet variances.
}
\description{
Creates the wavelet variance graph
}
\examples{
set.seed(999)
x=rnorm(100)
out = wvar(modwt(x))
autoplot( out )
}
\author{
JJB
}

