% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wvar.R
\name{autoplot.wvarcomp}
\alias{autoplot.wvarcomp}
\title{Compare Wavelet Variances Together}
\usage{
autoplot.wvarcomp(object, ...)
}
\arguments{
\item{object}{A \code{data.frame} containing both sets of variances.}

\item{...}{Not used...}
}
\value{
A ggplot2 graph containing both the wavelet variances.
}
\description{
Creates the wavelet variance graphs with classic and robust together
}
\examples{
set.seed(999)
x=rnorm(100)
Classic = wvar(modwt(x))
Robust = wvar(modwt(x), robust=TRUE)
compare.WV(Classic = Classic, Robust = Robust, split = FALSE)
}
\author{
JJB
}

