% Generated by roxygen2 (4.0.2): do not edit by hand
\name{avari_mo_arma}
\alias{avari_mo_arma}
\title{Compute Maximal-Overlap Allan Variance using Iterative Method}
\usage{
allan_variance(x)
}
\arguments{
\item{x}{A \code{vector} with dimensions N x 1.}
}
\value{
av A \code{list} that contains:
\itemize{
 \item{"clusters"}{The size of the cluster}
 \item{"allan"}{The allan variance}
 \item{"errors"}{The error associated with the variance calculation.}
}
}
\description{
Computation of Maximal-Overlap Allan Variance
}
\details{
Given \eqn{N} equally spaced samples with averaging time \eqn{\tau = n\tau _0}{tau = n*tau_0},
where \eqn{n} is an integer such that \eqn{ 1 \le n \le \frac{N}{2}}{1<= n <= N/2}.
Therefore, \eqn{n} is able to be selected from \eqn{\left\{ {n|n < \left\lfloor {{{\log }_2}\left( N \right)} \right\rfloor } \right\}}{{n|n< floor(log2(N))}}
Then, \eqn{M = N - 2n} samples exist.
The Maximal-overlap estimator is given as:
\deqn{\sigma _y^2\left( \tau  \right) = V\left( {M,\tau ,{\tau _0}} \right) = \frac{1}{{2{\tau ^2}M}}\sum\limits_{i = 0}^{M - 1} {{Z^2}\left( {i\tau_0 ,\tau } \right)}}
where \eqn{{Z^2}\left( {t\tau_0 ,\tau } \right) = {\left( {{X_t} - 2{X_{t + \tau }} + {X_{t + 2\tau }}} \right)^2}}, note \eqn{\tau_0 = 1}.
}
\examples{
set.seed(999)
# Simulate white noise (P 1) with sigma^2 = 4
N = 100000
white.noise = rnorm(N, 0, 2)
#plot(white.noise,ylab="Simulated white noise process",xlab="Time",type="o")
#Simulate random walk (P 4)
random.walk = cumsum(0.1*rnorm(N, 0, 2))
combined.ts = white.noise+random.walk
av_mat = avari_mo_arma(combined.ts)
}
\author{
JJB
}
\references{
Recipes for Degrees of Freedom of Frequency Stability Estimators, Charles A. Greenhall
}

