% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{boot_pval_gof}
\alias{boot_pval_gof}
\title{Generate the Confidence Interval for GOF Bootstrapped}
\usage{
boot_pval_gof(obj, obj_boot, B = 1000L, alpha = 0.05)
}
\arguments{
\item{alpha}{}

\item{theta}{}

\item{psi}{}
}
\value{
A \code{vec} that has the alpha/2.0 quantile and then the 1-alpha/2.0 quantile.
}
\description{
yaya
}

