% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bootstrap_gof_test}
\alias{bootstrap_gof_test}
\title{Compute the Bootstrapped GoF Test}
\usage{
bootstrap_gof_test(obj_value, bs_obj_values, alpha, bs_gof_p_ci)
}
\arguments{
\item{obj_value}{}

\item{alpha}{}

\item{bs_gof_p_ci}{A \code{bool} that indicates whether CIs should be included or not.}

\item{bs_obj_value}{}
}
\value{
A \code{vec} that has
\itemize{
\item Test Statistic
\item Low CI
\item Upper CI - BS
}
}
\description{
yaya
}

