% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{build_model_set}
\alias{build_model_set}
\title{Build List of Unique Models}
\usage{
build_model_set(combs, x)
}
\arguments{
\item{combs}{A \code{mat} that is a binary matrix (0,1) containing the combinations of different variables.}

\item{x}{A \code{vec<string>} that contains a list of model descriptors.}
}
\value{
A \code{set<string>} that contains the list of unique models.
}
\description{
Creates a set containing unique strings.
}
\keyword{internal}
