% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculate_psi_matrix}
\alias{calculate_psi_matrix}
\title{Calculate the Psi matrix}
\usage{
calculate_psi_matrix(D, v_hat, omega)
}
\arguments{
\item{D}{}

\item{v_hat}{}

\item{omega}{}
}
\value{
A \code{mat} that has the first column
}
\description{
Computes the Psi matrix using supplied parameters
}

