% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{checkEquality}
\alias{checkEquality}
\title{Check Equal Attributes in a GMWM List}
\usage{
checkEquality(obj.list)
}
\arguments{
\item{obj.list}{A list of \code{gmwm} objects.}
}
\value{
A list of square matrix indicating whether the objects have the same attributes.
}
\description{
Check whether the \code{gmwm} objects in \code{obj.list} have equal attributes.
}
\details{
This function is used in \code{\link{compare.models}}.
}
\author{
Wenchao
}
\keyword{internal}

