% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ci_eta3}
\alias{ci_eta3}
\title{Generate eta3 confidence interval}
\usage{
ci_eta3(y, dims, p)
}
\arguments{
\item{y}{A \code{vec} that computes the brickwalled modwt dot product of each wavelet coefficient divided by their length.}

\item{dims}{A \code{String} indicating the confidence interval being calculated.}

\item{p}{A \code{double} that indicates the \eqn{\left(1-p\right)*\alpha}{(1-p)*alpha} confidence level}
}
\value{
A \code{matrix} with the structure:
\itemize{
 \item{Column 1}{Wavelet Variance}
 \item{Column 2}{Chi-squared Lower Bounds}
 \item{Column 3}{Chi-squared Upper Bounds}
}
}
\description{
Computes the eta3 CI
}
\examples{
x=rnorm(100)
out = brick_wall(modwt_cpp(x), haar_filter())
sample = out[[1]]
y = sample*sample/length(sample)
ci_eta3(y, length(sample), 0.025)
}

