% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ci_eta3_robust}
\alias{ci_eta3_robust}
\title{Generate eta3 robust confidence interval}
\usage{
ci_eta3_robust(y, dims, p, eff)
}
\arguments{
\item{y}{A \code{vec} that computes the brickwalled modwt dot product of each wavelet coefficient divided by their length.}

\item{dims}{A \code{String} indicating the confidence interval being calculated.}

\item{p}{A \code{double} that indicates the \eqn{\left(1-p\right)*\alpha}{(1-p)*alpha} confidence level}

\item{eff}{A \code{double} that indicates the efficiency.}
}
\value{
A \code{matrix} with the structure:
\itemize{
 \item{Column 1}{Robust Wavelet Variance}
 \item{Column 2}{Chi-squared Lower Bounds}
 \item{Column 3}{Chi-squared Upper Bounds}
}
}
\description{
Computes the eta3 robust CI
}
\examples{
x = rnorm(100)
# Uses the internal MODWT function not associated with an S3 class.
decomp = modwt_cpp(x, "haar", 4, boundary="periodic")
signal_modwt_bw = brick_wall(decomp, select_filter("haar"), "modwt")
y = wave_variance(signal_modwt_bw, robust = TRUE,  eff = 0.6)
ci_wave_variance(signal_modwt_bw, y, type = "eta3", p = 0.025, robust = TRUE, eff = 0.6)
}

