% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection_methods.R
\name{comb.mat}
\alias{comb.mat}
\title{Create Combination Matrix}
\usage{
comb.mat(n)
}
\arguments{
\item{n}{The number of variables.}
}
\value{
Returns a binary matrix (e.g. 1 or 0) entries
}
\description{
Generates a matrix containing all the combination listings
}
\details{
Port expand.grid to C++ at a later time...
}
\keyword{internal}

