% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GMWM.R
\name{compare.models}
\alias{compare.models}
\title{Graphically Compare GMWM Model Fit}
\usage{
compare.models(..., background = "white", split = TRUE, CI = TRUE,
  auto.label.wvar = T, transparence = 0.1, line.color = NULL,
  CI.color = NULL, line.type = NULL, point.size = NULL,
  point.shape = NULL, title = "Haar Wavelet Variance Representation",
  title.size = 15, axis.label.size = 13, axis.tick.size = 11,
  axis.x.label = expression(paste("Scale ", tau)),
  axis.y.label = expression(paste("Wavelet Variance ", nu)),
  facet.label.size = 13, legend.label = NULL, legend.title = "",
  legend.key.size = 1.3, legend.title.size = 13, legend.text.size = 13,
  nrow = 1)
}
\arguments{
\item{...}{Several \code{gmwm} objects}

\item{split}{A \code{boolean} indicating true or false to place model fits on different or the same graphs.}
}
\value{
A ggplot2 panel containing two graphs of the wavelet variance.
}
\description{
Creates GMWM model fits of different models on split graphs within the same panel.
}
\examples{
\dontrun{# AR
set.seed(8836)
n = 200
x = gen.ts(AR1(phi = .1, sigma2 = 1) + AR1(phi = 0.95, sigma2 = .1), n)
GMWM1 = gmwm(AR1(), data = x)
GMWM2 = gmwm(2*AR1(), data = x)
compare.models(GMWM1, GMWM2, split = FALSE)}
}
\author{
JJB, Wenchao
}

