% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wvar.R
\name{compare.wvar.imu}
\alias{compare.wvar.imu}
\title{Compare Wavelet Variances for \code{imu} Object}
\usage{
compare.wvar.imu(obj.list, object.names, background = "white", CI = TRUE,
  auto.label.wvar = T, transparence = 0.1, line.color = NULL,
  CI.color = NULL, line.type = NULL, point.size = NULL,
  point.shape = NULL, title = "Haar Wavelet Variance Representation",
  title.size = 15, axis.label.size = 13, axis.tick.size = 11,
  axis.x.label = expression(paste("Scale ", tau)),
  axis.y.label = expression(paste("Wavelet Variance ", nu)),
  facet.label.size = 13, facet.label.background = "#003C7D33",
  legend.label = NULL, legend.title = "", legend.key.size = 1.3,
  legend.title.size = 13, legend.text.size = 13)
}
\arguments{
\item{obj.list}{A \code{list} of \code{wvar.imu} objects.}

\item{object.names}{A \code{character vector} that contains the names of objects to be compared.}

\item{background}{A \code{string} that determines the graph background. It can be \code{'grey'} or \code{'white'}.}

\item{CI}{A \code{boolean} that indicates whether the confidence interval should be plotted.}

\item{auto.label.wvar}{A \code{boolean} that indicates whether legend label should indicate the objects are robust or classical.}

\item{transparence}{A \code{double} that ranges from 0 to 1 that controls the transparency of the graph.}

\item{line.color}{A \code{vector} of \code{string} that indicates the color of lines.}

\item{CI.color}{A \code{vector} of \code{string} that indicates the color of confidence interval.}

\item{line.type}{A \code{vector} of \code{string} that indicates the type of lines.}

\item{point.size}{A \code{vector} of \code{integer} that indicates the size of point.}

\item{point.shape}{A \code{vector} of \code{integer} that indicates the shape of point.}

\item{title}{A \code{string} that indicates the title of the graph.}

\item{title.size}{An \code{integer} that indicates the size of title.}

\item{axis.label.size}{An \code{integer} that indicates the size of label.}

\item{axis.tick.size}{An \code{integer} that indicates the size of tick mark.}

\item{axis.x.label}{A \code{string} that indicates the label on x axis.}

\item{axis.y.label}{A \code{string} that indicates the label on y axis.}

\item{facet.label.size}{An \code{integer} that indicates the size of facet label.}

\item{facet.label.background}{A \code{string} that indicates the background color of the facet label.}

\item{legend.label}{A \code{vector} of \code{string} that indicates the labels on legend.}

\item{legend.title}{A \code{string} that indicates the title of legend.}

\item{legend.key.size}{A \code{double} that indicates the size of key (in centermeters) on legend.}

\item{legend.title.size}{An \code{integer} that indicates the size of title on legend.}

\item{legend.text.size}{An \code{integer} that indicates the size of key label on legend.}
}
\description{
Internal Function to compare the wavelet variance for \code{imu} object.
}
\author{
Wenchao
}
\seealso{
\code{\link{compare.wvar}}
}

