% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMWM.R
\name{compare_eff}
\alias{compare_eff}
\title{Graphically Compare Classical with Robust GMWM Models}
\usage{
compare_eff(
  ...,
  display.eff = T,
  order.eff = T,
  facet.label = NULL,
  background = "white",
  transparence = 0.05,
  CI.color = NULL,
  line.color = NULL,
  line.type = NULL,
  point.size = NULL,
  point.shape = NULL,
  title = NULL,
  title.size = 18,
  axis.label.size = 16,
  axis.tick.size = 11,
  facet.label.size = 13,
  facet.label.background = "#003C7D33",
  axis.x.label = expression(paste("Scale ", tau)),
  axis.y.label = expression(paste("Wavelet Variance ", nu))
)
}
\arguments{
\item{...}{Several \code{gmwm} objects, and they must be constrcuted by the same data and same model. Or one \code{rgmwm} object created by \code{\link{rgmwm}} function.}

\item{display.eff}{A \code{boolean} indicating whether the classical/robust method with efficiency should be displayed in the facet label. It is only used when \code{facet.label} is NULL.}

\item{order.eff}{A \code{boolean} indicating whether the object should be ordered by their efficiency value.}

\item{facet.label}{A \code{character vector} indicating what should be displayed in the facet labels.}

\item{background}{A \code{string} that determines the graph background. It can be \code{'grey'} or \code{'white'}.}

\item{transparence}{A \code{double} between 0 to 1 that controls the transparency of confidence interval.}

\item{CI.color}{A \code{character vector} that indicates the color of the confidence interval (e.g. 'black', 'red', '#003C7D', etc.)}

\item{line.color}{A \code{character vector} that indicates the color of lines.}

\item{line.type}{A \code{character vector} that indicates the type of lines.}

\item{point.size}{A \code{integer vector} that indicates the size of points on lines.}

\item{point.shape}{A \code{integer vector} that indicates the shape of points on lines.}

\item{title}{A \code{string} that indicates the title of the graph.}

\item{title.size}{An \code{integer} that indicates the size of title.}

\item{axis.label.size}{An \code{integer} that indicates the size of label.}

\item{axis.tick.size}{An \code{integer} that indicates the size of tick mark.}

\item{facet.label.size}{An \code{integer} that indicates the size of facet label.}

\item{facet.label.background}{A \code{string} that indicates the background color of the facet label.}

\item{axis.x.label}{A \code{string} that indicates the label on x axis.}

\item{axis.y.label}{A \code{string} that indicates the label on y axis.}
}
\description{
Creates a table of graphs to compare GMWM models constructed by classical and robust methods.
}
\section{Constraints on \code{gmwm} objects}{

This function is designed to compare the \code{gmwm} objects with different computation methods and efficiency values. Therefore,
\code{gmwm} objects supplied to this function are assumed to have different efficiency values and are constructed by 
the same data and same model. The function will check the constraint before generating the graph.
}

\section{How to change \code{CI.color} and \code{facet.label}}{

To change \code{CI.color} and \code{facet.label}, the user must supply a vector. 
If the user compares \code{N} {gmwm} objects, then he/she must supply a vector of length 
\code{N} to \code{CI.color} and \code{facet.label}. Please check examples for help. Additionally,
\code{order.eff} will change the order how objects are plotted. User is recommended to plot the graph
without changing the aesthetics firstly. Then, the user can generate the graph by supplying elements to graphical parameters, e.g. \code{CI.color},
\code{facet.label}, in left-to-right order.
}

\section{How to change other graphical aesthetics}{

\code{line.type}, \code{line.color}, \code{point.size}, \code{point.size} must be a \code{vector}. 
If the user compares \code{N} \code{gmwm} objects, then a vector of length \eqn{4 \times N}{ 4 x N } must be supplied to these 
parameters. For example, if the user wants to compare \code{gmwm1} with \code{gmwm2}, the order to change the 
graphical aesthetics is:
"gmwm1 Empirical WV, gmwm2 Empirical WV, gmwm1 lower bound of CI, gmwm2 lower bound of CI, gmwm1 higher bound of 
CI, gmwm2 higher bound of CI, gmwm1 Implied WV, gmwm2 Implied WV."

Please check examples for help.
}

\section{Other details}{


\code{melt()} is used to convert the data frame from wide format to long format. However,
\code{melt()} cannot deal with expressions, so the user cannot supply expressions to the parameter 
\code{facet.label}. Users should convert expressions into characters via \code{as.character()}.

If you meet the error "polygon edge not found", it is complaining that you don't have enough space to
plot the graph. If you are using RStudio, you can adjust the plot window. You can also open graphical window by using
\code{quartz()} (Mac/Linux) or \code{windows()} (Windows PC).
}

\examples{
\dontrun{
#Simulate data
set.seed(8836)
n = 1000
x = gen_gts(n, AR1(phi = .1, sigma2 = 1) + AR1(phi = 0.95, sigma2 = .1))
x[1] = 1000 #Robust gmwm works better for contaminated data
GMWM1 = gmwm(2*AR1()+RW(), data = x, robust  = TRUE, eff = 0.9)
GMWM2 = gmwm(2*AR1()+RW(), data = x, robust  = TRUE, eff = 0.6)
GMWM3 = gmwm(2*AR1()+RW(), data = x, robust  = FALSE)

# How order.eff works:
compare_eff(GMWM1, GMWM2, GMWM3, order.eff = FALSE)
compare_eff(GMWM1, GMWM2, GMWM3, order.eff = TRUE) 
# order.eff=TRUE: The objects are plotted in descending order of model efficiency.


# How to modify facet.label
compare_eff(GMWM2, GMWM3, order.eff = FALSE, facet.label = c('Rob. eff. 0.6', 'Cla.') )
compare_eff(GMWM2, GMWM3, order.eff = TRUE, facet.label = c('Cla.', 'Rob. eff. 0.6') )

# How to modify graph aesthetics
compare_eff(GMWM2, GMWM3, order.eff = FALSE, CI.color = c("#003C7D", "#F47F24"),
           line.color = rep(c("#003C7D", "#F47F24"), 4) )
compare_eff(GMWM2, GMWM3, order.eff = TRUE, CI.color = c("#F47F24", "#003C7D"),
           line.color = rep(c("#F47F24", "#003C7D"), 4) )
}
}
\author{
Wenchao
}
