% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wvar.R
\name{compare_wvar}
\alias{compare_wvar}
\title{Compare Wavelet Variances}
\usage{
compare_wvar(..., background = "white", split = TRUE, CI = TRUE,
  auto.label.wvar = T, transparence = 0.1, line.color = NULL,
  CI.color = NULL, line.type = NULL, point.size = NULL,
  point.shape = NULL, title = "Haar Wavelet Variance Representation",
  title.size = 15, axis.label.size = 13, axis.tick.size = 11,
  axis.x.label = expression(paste("Scale ", tau)),
  axis.y.label = expression(paste("Wavelet Variance ", nu)),
  units = c(bquote(rad^2/s^2), bquote(m^2/s^4)), facet.label.size = 13,
  facet.label.background = "#003C7D33", legend.label = NULL,
  legend.title = "", legend.key.size = 1.3, legend.title.size = 13,
  legend.text.size = 13, nrow = 1)
}
\arguments{
\item{...}{Any number of \code{wvar} or \code{wvar.imu} objects.}

\item{background}{A \code{string} that determines the graph background. It can be \code{'grey'} or \code{'white'}.}

\item{split}{A \code{boolean} that indicates whether the graphs should be separate (TRUE) or graphed ontop of each other (FALSE)}

\item{CI}{A \code{boolean} that indicates whether the confidence interval should be plotted.}

\item{auto.label.wvar}{A \code{boolean} that indicates whether legend label should indicate the objects are robust or classical.}

\item{transparence}{A \code{double} that ranges from 0 to 1 that controls the transparency of the graph}

\item{line.color}{A \code{vector} of \code{string} that indicates the color of lines.}

\item{CI.color}{A \code{vector} of \code{string} that indicates the color of confidence interval.}

\item{line.type}{A \code{vector} of \code{string} that indicates the type of lines.}

\item{point.size}{A \code{vector} of \code{integer} that indicates the size of point.}

\item{point.shape}{A \code{vector} of \code{integer} that indicates the shape of point.}

\item{title}{A \code{string} that indicates the title of the graph.}

\item{title.size}{An \code{integer} that indicates the size of title.}

\item{axis.label.size}{An \code{integer} that indicates the size of label}

\item{axis.tick.size}{An \code{integer} that indicates the size of tick mark}

\item{axis.x.label}{A \code{string} that indicates the label on x axis}

\item{axis.y.label}{A \code{string} that indicates the label on y axis}

\item{units}{A two-element vector indicating the units of gyroscope and accelerometer sensor. Set it to \code{NULL} if units are not needed.}

\item{facet.label.size}{An \code{integer} that indicates the size of facet label}

\item{facet.label.background}{A \code{string} that indicates the background color of the facet label}

\item{legend.label}{A \code{vector} of \code{string} that indicates the labels on legend.}

\item{legend.title}{A \code{string} that indicates the title of legend}

\item{legend.key.size}{A \code{double} that indicates the size of key (in centermeters) on legend}

\item{legend.title.size}{An \code{integer} that indicates the size of title on legend}

\item{legend.text.size}{An \code{integer} that indicates the size of key label on legend}

\item{nrow}{An \code{integer} that indicates number of rows}
}
\description{
Compare the estimates given by the classical and robust methods of 
calculating the wavelet variance.
}
\note{
If you meet the error "polygon edge not found", RStudio is complaining that you don't have enough 
space to plot the graph. You can adjust the graphics window, or open an external window. The 
function \code{\link{external_graphs}} can be used and it works for all operating systems.

When \code{wvar.imu} objects are supplied, some parameters, e.g. \code{split} and \code{nrow}, are 
invalid, since the graph is plot seperately and put in 2 rows by default.
}
\examples{
\dontrun{
## Case1: Supplied objects are \\code{wvar}:
data1 = gen_gts(1000, AR1(phi = .32, sigma2=.01))
data2 = gen_gts(2000, ARMA(ar=c(.8,.1), ma=c(.3), sigma2=1))
data3 = gen_gts(4000, AR1(phi = .32, sigma2=1))

wv1 = wvar(data1, robust = TRUE)
wv2 = wvar(data2)
wv3 = wvar(data3)

compare_wvar(wv1, wv2)
compare_wvar(wv1, wv2, CI = FALSE)
compare_wvar(wv1, wv2, split = FALSE)
compare_wvar(wv1, wv2, wv3, split = FALSE)

# Change default setting
color = c('green','red','blue')
label = c('1','2','3')
compare_wvar(wv1, wv2, wv3, 
             line.color = color, CI.color = color,
              legend.label = label)
compare_wvar(wv1, wv2, wv3, 
             line.color = color, CI.color = color, 
             legend.label = label, split = FALSE)

## Case2: Supplied objects are \\code{wvar.imu}:
if(!require("imudata")){
   install_imudata()
   library("imudata")
}

data(imu6)
test1 = imu(imu6, gyros = 1:3, accels = 4:6, axis = c('X', 'Y', 'Z'), freq = 100)
wv1 = wvar(test1)

test2 = imu(imu6, gyros = 1, accels = 3:4, axis = c('X','X','Y'), freq = 100)
wv2 = wvar(test2, robust = T)

compare_wvar(wv1, wv2)
compare_wvar(wv1, wv2, auto.label.wvar = F, legend.label = c('data1', 'data2'))
}
}
\author{
JJB, Wenchao
}
