% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\docType{methods}
\name{count_models_alpha}
\alias{count_models_alpha}
\title{Count Number of Models (Alphanumeric)}
\arguments{
\item{desc}{A \code{vector<string>} that contains the model type}
}
\value{
A \code{vec} with the model counts.
}
\description{
Return a model count
}
\details{
The types of models supported are:
\itemize{
  \item{"AR1"}{a first order autoregressive process with parameters \eqn{(\phi,\sigma^2)}{phi, sigma^2}}
  \item{"DR"}{a drift with parameter \eqn{\omega}{omega}}
  \item{"QN"}{a quantization noise process with parameter \eqn{Q}}
  \item{"RW"}{a random walk process with parameter \eqn{\sigma^2}{sigma^2}}
  \item{"WN"}{a white noise process with parameter \eqn{\sigma^2}{sigma^2}}
}
Note: This order is due to the fact that the models are created by inserting into a C++ Map. The map sorts these values...
This function is NOT available in R.
}

