% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{count_models_numeric}
\alias{count_models_numeric}
\title{Count Number of Model instances}
\usage{
count_models_numeric(desc)
}
\arguments{
\item{desc}{A \code{vec} that contains the model type}
}
\value{
A \code{vec} with the model counts.
}
\description{
Return a model count
}
\details{
Hard coded value for the number of models we currently support.
The types of models supported are:
\itemize{
  \item{0}{AR1}
  \item{1}{Drift}
  \item{2}{White Noise}
  \item{3}{Random Walk}
  \item{4}{Q Noise}
}
}
\examples{
count_models(1,2,3,4,4,0,0,0,3,4)
}

