% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection_methods.R
\name{cust.model.score}
\alias{cust.model.score}
\title{Formats the model score matrix}
\usage{
cust.model.score(out, model.names)
}
\arguments{
\item{out}{A \code{list} containing the model matrix}

\item{model.names}{A \code{character vector} that contains a list of the model names.}
}
\value{
An updated matrix in the ith position of the list.
}
\description{
The model score matrix receives the appropriate model numbering, col descriptors, and ordering of models.
}
\keyword{internal}

