% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{D_matrix}
\alias{D_matrix}
\title{D Matrix}
\usage{
D_matrix(At_j, omega, diff)
}
\arguments{
\item{At_j}{A \code{mat} that is of dimensions J x P containing the sederivative of A(theta_hat_j)/dTheta_hat_(j,tau)}

\item{omega}{A \code{mat} that is of dimension P x P used in obtaining the GMWM estimator.}

\item{diff}{A \code{vec} that is the difference of the WV empirical and WV theoretical}
}
\value{
A \code{mat}
}
\description{
D Matrix
}
\details{
TBA
}

