% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{decomp_theoretical_wv}
\alias{decomp_theoretical_wv}
\title{Each Models Process Decomposed to WV}
\usage{
decomp_theoretical_wv(theta, desc, objdesc, tau)
}
\arguments{
\item{theta}{A \code{vec} containing the list of estimated parameters.}

\item{desc}{A \code{vector<string>} containing a list of descriptors.}

\item{objdesc}{A \code{field<vec>} containing a list of object descriptors.}

\item{tau}{A \code{vec} containing the scales e.g. 2^(1:J)}
}
\value{
A \code{mat} containing the wavelet variance of each process in the model
}
\description{
This function computes each process to WV (haar) in a given model.
}
\examples{
x.sim = gen_ar1( N = 10000, phi = 0.9, sigma2 = 4 )
ntau = floor(log(length(x.sim),2))
tau = 2^(1:ntau)
wv.theo = ar1_to_wv(phi = 0.9, sig2 = 16, tau)
plot(tau, wv.theo, col = "red")
}
\keyword{internal}

