% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{decomp_to_theo_wv}
\alias{decomp_to_theo_wv}
\title{Decomposed WV to Single WV}
\usage{
decomp_to_theo_wv(decomp)
}
\arguments{
\item{decomp}{A \code{mat} with scales as rows and processes as columns}
}
\value{
A \code{vec} containing the wavelet variance of the process for the overall model
}
\description{
This function computes the combined processes to WV (haar) in a given model.
}
\examples{
x.sim = gen_ar1( N = 10000, phi = 0.9, sigma2 = 4 )
ntau = floor(log(length(x.sim),2))
tau = 2^(1:ntau)
wv.theo = ar1_to_wv(phi = 0.9, sig2 = 16, tau)
plot(tau, wv.theo, col = "red")
}
\keyword{internal}

