% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{deriv_2nd_ar1}
\alias{deriv_2nd_ar1}
\title{Analytic second derivative matrix for AR(1) process}
\usage{
deriv_2nd_ar1(phi, sigma2, tau)
}
\arguments{
\item{phi}{A \code{double} corresponding to the phi coefficient of an AR(1) process.}

\item{sigma2}{A \code{double} corresponding to the error term of an AR(1) process.}

\item{tau}{A \code{vec} containing the scales e.g. \eqn{2^{\tau}}{2^tau}}
}
\value{
A \code{matrix} with the first column containing the
 second partial derivative with respect to \eqn{\phi}{phi} and
  the second column contains the second partial derivative with 
  respect to \eqn{\sigma ^2}{sigma^2}
}
\description{
Calculates the second derivative for the AR(1) process and places it into a matrix form.
The matrix form in this case is for convenience of the calculation.
}
\examples{
deriv_2nd_ar1(.3, 1, 2^(1:5))
}
\author{
James Joseph Balamuta (JJB)
}

