% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{deriv_2nd_ma1}
\alias{deriv_2nd_ma1}
\title{Analytic second derivative for MA(1) process}
\usage{
deriv_2nd_ma1(theta, sig2, tau)
}
\arguments{
\item{theta}{A \code{double} corresponding to the phi coefficient of an MA(1) process.}

\item{sig2}{A \code{double} corresponding to the error term of an MA(1) process.}

\item{tau}{A \code{vec} that contains the scales to be processed (e.g. 2^(1:J))}
}
\value{
A \code{matrix} with the first column containing the second partial derivative with respect to \eqn{\theta}{\theta},
 the second column contains the partial derivative with respect to \eqn{\theta}{\theta} and \eqn{\sigma ^2}{sigma^2},
 and lastly we have the second partial derivative with respect to \eqn{\sigma ^2}{sigma^2}.
}
\description{
To easy calculation, we assume a matrix structure.
}
\details{
See the supporting vignette documentation.
}
\examples{
deriv_2nd_ma1(.3, 1, 2^(1:5))
}
\author{
JJB
}

