% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{deriv_ar1}
\alias{deriv_ar1}
\title{Analytic D matrix for AR(1) process}
\usage{
deriv_ar1(phi, sig2, tau)
}
\arguments{
\item{phi}{A \code{double} corresponding to the phi coefficient of an AR(1) process.}

\item{sig2}{A \code{double} corresponding to the error term of an AR(1) process.}

\item{tau}{A \code{vec} that contains the scales to be processed (e.g. 2^(1:J))}
}
\value{
A \code{matrix} with the first column containing the partial derivative with respect to \eqn{\phi ^2}{sigma^2} and the second column contains the partial derivative with respect to \eqn{\sigma ^2}{sigma^2}
}
\description{
Obtain the first derivative of the AR(1) process.
}
\details{
See the supporting vignette documentation.
}
\examples{
deriv_ar1(.3, 1, 2^(1:5))
}
\author{
JJB
}

