% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.model.R
\name{desc.to.ts.model}
\alias{desc.to.ts.model}
\title{Create a ts.model from desc string}
\usage{
desc.to.ts.model(desc)
}
\arguments{
\item{desc}{A \code{character} vector containing: \code{"AR1"},\code{"DR"},\code{"WN"},\code{"RW"},\code{"QN"}}
}
\value{
An S3 object with called ts.model with the following structure:
\itemize{
 \item{desc}
 \item{theta}
}
}
\description{
Sets up the necessary backend for using Cpp functions to build R ts.model objects
}
\examples{
desc.to.ts.model(c("AR1","WN"))
}
\author{
JJB
}
\keyword{internal}

