% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{dim.gts}
\alias{dim.gts}
\alias{dim.imu}
\alias{dim.lts}
\title{Dimensions of GMWM Objects}
\usage{
\method{dim}{gts}(x)

\method{dim}{imu}(x)

\method{dim}{lts}(x)
}
\arguments{
\item{x}{A \code{gts}, \code{imu}, \code{lts} object.}
}
\value{
A \code{vector} with two elements.

The first element contains the number of rows. 
The second element contains the number of columns.
}
\description{
Returns the dimensions of a 
\code{gts}, \code{imu}, or \code{lts} object.
}
\author{
JJB
}

