% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_install.R
\name{dl.file}
\alias{dl.file}
\title{Better file downloader}
\usage{
dl.file(url, ...)
}
\arguments{
\item{url}{A \code{string} that contains a url}

\item{...}{Additional parameters to be passed to \code{\link[utils]{download.file}}}
}
\value{
A downloaded file...
}
\description{
Replaces download.file's auto detect mechanism
}
\details{
See \url{https://support.rstudio.com/hc/en-us/articles/206827897-Secure-Package-Downloads-for-R}
for information on downloading via HTTPS. Also, read through the changes in \code{\link[utils]{download.file}}.
}
\seealso{
\code{\link[utils]{download.file}}
}
\keyword{internal}

