% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{do_polyroot_cpp}
\alias{do_polyroot_cpp}
\title{Root Finding C++}
\usage{
do_polyroot_cpp(z)
}
\arguments{
\item{z}{A \code{vec<complex<double>} (complex vector) that has 1 in the beginning (e.g. c(1,3i,-3i))}
}
\description{
Vroom Vroom
}
\keyword{internal}
