% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.acf}
\alias{.acf}
\title{Auto-Covariance and Correlation Functions}
\usage{
.acf(x, lagmax = 0L, cor = TRUE, demean = TRUE)
}
\arguments{
\item{x}{A \code{matrix} with dimensions \eqn{N \times S}{N x S} or N observations and S processes}

\item{lagmax}{A \code{integer}}

\item{cor}{A \code{bool} indicating whether the correlation 
(\code{TRUE}) or covariance (\code{FALSE}) should be computed.}

\item{demean}{A \code{bool} indicating whether the data should be detrended
(\code{TRUE}) or not (\code{FALSE})}
}
\description{
The acf function computes the estimated
autocovariance or autocorrelation for both univariate and multivariate cases.
}
\keyword{internal}
