% Generated by roxygen2 (4.0.2): do not edit by hand
\name{DR}
\alias{DR}
\title{Create an Drift (DR) Process}
\usage{
DR(slope = 5)
}
\arguments{
\item{slope}{A \code{double} value for the slope of a DR process.}
}
\value{
An S3 object with called ts.model with the following structure:
\itemize{
 \item{desc}{"DR"}
 \item{theta}{slope}
}
}
\description{
Sets up the necessary backend for the DR process.
}
\examples{
DR()
DR(slope=3.4)
}
\author{
JJB
}

