% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{e_drift}
\alias{e_drift}
\title{Expected value DR}
\usage{
e_drift(omega, n_ts)
}
\arguments{
\item{omega}{A \code{double} corresponding to variance of drift.}

\item{n_ts}{An \code{int} indicating the length of the time series.}
}
\value{
A \code{vec} containing the expected value of the drift.
}
\description{
This function computes the expected value of a drift process.
}
\examples{
e_drift(1,200)
}
\keyword{internal}
