% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\docType{methods}
\name{expand_grid_red}
\alias{expand_grid_red}
\title{Expand Grid for Same Dimensional Case}
\usage{
expand_grid_red(nx)
}
\arguments{
\item{nx}{An \code{integer} of length f that contains the initial values of the time series in reverse.}
}
\value{
x A \code{matrix} listing values from 1...nx in one column and 1...1, 2...2,....,n...n, in the other
}
\description{
Creates the different pairings possible with two different variables.
}
\details{
This function is hidden and is not accessible from R.
}
\author{
R Core Team and JJB
}

