% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expand_sarima}
\alias{expand_sarima}
\title{Expand Parameters for an SARIMA object}
\usage{
expand_sarima(params, objdesc)
}
\arguments{
\item{params}{A \code{vec} containing the theta values of the parameters.}

\item{objdesc}{A \code{vec} containing the model term information.}
}
\value{
A \code{field<vec>} of size two as follows:
\itemize{
  \item AR    values
  \item THETA values
}
}
\description{
Creates an expanded PHI and THETA vector for use in other objects.
}
\details{
The \code{objdesc} is assumed to have the structure of:
\itemize{
\item AR(p)
\item MA(q)
\item SAR(P)
\item SMA(Q)
\item Seasons
}
}
\examples{
m = expand_sarima(c(0.5,.2,0,.1,.92,.83,.42,.33,.12), c(2,2,2,3,12))
}

