% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_cov_cpp}
\alias{fast_cov_cpp}
\title{Computes the (MODWT) wavelet covariance matrix using Chi-square confidence interval bounds}
\usage{
fast_cov_cpp(ci_hi, ci_lo)
}
\arguments{
\item{ci_hi}{A \code{vec} that contains the upper confidence interval points.}

\item{ci_lo}{A \code{vec} that contains the lower confidence interval points.}
}
\value{
A diagonal matrix.
}
\description{
Calculates the (MODWT) wavelet covariance matrix using Chi-square confidence interval bounds
}
\examples{
\dontrun{
x=runif(100)
y=x+3
fast_cov_cpp(y,x)
}
}
\keyword{internal}
