% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_biwc}
\alias{find_biwc}
\title{Obtain Tuning Constant crob.bw}
\usage{
find_biwc(eff)
}
\arguments{
\item{eff}{A \code{double} that represents the desired level of efficiency as compared to the classic MLE.}
}
\value{
A \code{double} that contains the optimized biwc.
}
\description{
Objective function that finds tuning constant
}
\examples{
find_biwc(0.6)
}

