% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{format_ci}
\alias{format_ci}
\title{Format the Confidence Interval for Estimates}
\usage{
format_ci(theta, se, alpha)
}
\arguments{
\item{theta}{}

\item{se}{}

\item{alpha}{}
}
\value{
A \code{mat} that has:
\itemize{
\item Column 1: Lo CI
\item Column 2: Hi CI
\item Column 3: SE
}
}
\description{
Creates hi and lo confidence based on SE and alpha.
}

