% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gen.ts.R
\name{gen.ts}
\alias{gen.ts}
\title{Generate Time Series based on Model}
\usage{
gen.ts(model, N = 1000)
}
\arguments{
\item{model}{A \code{ts.model} or \code{gmwm} object containing one of the allowed models.}

\item{N}{An \code{interger} containing the amount of observations for the time series.}
}
\value{
A \code{vec} that contains combined time series.
}
\description{
Create a time series based on a supplied time series model.
}
\details{
This function accepts either a ts.model object (e.g. AR1(phi = .3, sigma2 =1) + WN(sigma2 = 1)) or a gmwm object.
}
\examples{
# AR
set.seed(1336)
model = AR1(phi = .99, sigma = 1) + WN(sigma2=1)
gen.ts(model)
}

