% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/gen_process.cpp, src/gen_process.h
\name{gen_arma}
\alias{gen_arma}
\title{Generate ARMA}
\usage{
gen_arma(N, ar, ma, sigma2 = 1.5, n_start = 0L)
}
\arguments{
\item{N}{An \code{integer} for signal length.}

\item{ar}{A \code{vec} that contains the AR coefficients.}

\item{ma}{A \code{vec} that contains the MA coefficients.}

\item{sigma2}{A \code{double} that contains process variance.}

\item{n_start}{An \code{unsigned int} that indicates the amount of observations to be used for the burn in period.}
}
\value{
A \code{vec} that contains the generated observations.
}
\description{
Generate observations for a supplied ARMA model.
}
\details{
The innovations are generated from a normal distribution.
}
\examples{
gen_arma(100, c(.3,.5), c(.1), 1, 0)
}
\keyword{internal}

