% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/gen_process.cpp, src/gen_process.h
\name{gen_arma11}
\alias{gen_arma11}
\title{Generate an ARMA(1,1) sequence}
\usage{
gen_arma11(N, phi = 0.1, theta = 0.3, sigma2 = 1)
}
\arguments{
\item{N}{An \code{integer} for signal length.}

\item{phi}{A \code{double} that contains autoregressive.}

\item{theta}{A \code{double} that contains moving average.}

\item{sigma2}{A \code{double} that contains process variance.}
}
\value{
A \code{vec} containing the MA(1) process.
}
\description{
Generate an ARMA(1,1) sequence given \eqn{\phi}, \eqn{\theta}, and \eqn{\sigma^2}.
}
\details{
The function implements a way to generate the \eqn{x_t}{x[t]} values without calling the general ARMA function.
The autoregressive order 1 and moving average order 1 process is defined as \eqn{{x_t} = {\phi _1}{x_{t - 1}} + {w_t} + {\theta _1}{w_{t - 1}} }{x[t] = phi*x[t-1] + w[t] + theta*w[t-1]},
 where \eqn{{w_t}\mathop  \sim \limits^{iid} N\left( {0,\sigma _w^2} \right)}{w[t] ~ N(0,sigma^2) iid}

The function first generates a vector of white noise using \code{\link[gmwm]{gen_wn}} and then obtains the
ARMA values under the above equation.
}
\examples{
gen_ma1(10, .2, 1.2)
}
\keyword{internal}

