% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gen_dr}
\alias{gen_dr}
\title{Generate a drift}
\usage{
gen_dr(N, slope)
}
\arguments{
\item{N}{An \code{integer} for signal length.}

\item{slope}{A \code{double} that contains drift slope}
}
\value{
gd A \code{vec} containing the drift.
}
\description{
Generates a drift sequence with a given slope.
}
\examples{
gen_dr(10, 8.2)
}

