% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gen_wn}
\alias{gen_wn}
\title{Generate a white noise process}
\usage{
gen_wn(N, sigma_WN)
}
\arguments{
\item{N}{An \code{integer} for signal length.}

\item{sigma_WN}{A \code{double} that contains process standard deviation.}
}
\value{
wn A \code{vec} containing the white noise.
}
\description{
Generates a white noise process with variance parameter sigma.
}
\examples{
gen_wn(10, 1.5)
}

