% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getColors}
\alias{getColors}
\title{Get N Colors}
\usage{
getColors(palette, n, rm = NULL)
}
\arguments{
\item{palette}{A \code{string} that indicates the name of the palette.}

\item{n}{An \code{integer} that specifies the number of colors to be generated.}

\item{rm}{An \code{integer vector} that specifies the index of colors to be removed.}
}
\description{
Creates n colors from specific palette
}
\details{
Currently, only the palette 'Set1' is implemented in this function.

'Set1' palette in the package \code{RColorBrewer} originally has 9 colors, but we delete the 
yellow color (the 6th one) since it is too much vibrant. 

If \code{n} is larger than the mamixum number of colors in the palette, the function
will repeat the palette until \code{n} colors is created.
}
\examples{
getColors('Set1', 10)
}
\keyword{internal}

