% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getEff}
\alias{getEff}
\title{Get \code{gmwm} Efficiency Values}
\usage{
getEff(obj.list)
}
\arguments{
\item{obj.list}{A \code{list} of \code{gmwm} object}
}
\value{
A \code{numeric vector}. If the object is computed by classical method, it will return 1.1.
}
\description{
Get efficiency values from a list of \code{gmwm} object
}
\examples{
set.seed(8836)
n = 1000
x = gen.gts(AR1(phi = .1, sigma2 = 1) + AR1(phi = 0.95, sigma2 = .1), n)
GMWM1 = gmwm(2*AR1()+RW(), data = x, robust  = FALSE)
GMWM2 = gmwm(2*AR1()+RW(), data = x, robust  = TRUE, eff = 0.1)
GMWM3 = gmwm(2*AR1()+RW(), data = x, robust  = TRUE, eff = 0.6)
obj.list = list(GMWM1, GMWM2, GMWM3)
getEff(obj.list)
}
\keyword{internal}

