% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{getModel.gmwm}
\alias{getModel.gmwm}
\title{Get the model in a \code{gmwm} object}
\usage{
getModel.gmwm(object)
}
\arguments{
\item{object}{A \code{gmwm} object}
}
\value{
A \code{string} containing the model
}
\description{
Extracts and formats the model string.
}
\keyword{internal}

