% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/options.R
\name{ggColor}
\alias{ggColor}
\title{Emulate ggplot2 default color palette}
\usage{
ggColor(n)
}
\value{
A \code{vector} containing \code{n} colors
}
\description{
Emulate ggplot2 default color palette
}
\author{
John Colby
}

