% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{conv.ar1.to.gm}
\alias{conv.ar1.to.gm}
\alias{conv.gm.to.ar1}
\title{GM Conversion}
\usage{
conv.ar1.to.gm(theta, process.desc, freq)

conv.gm.to.ar1(theta, process.desc, freq)
}
\arguments{
\item{theta}{A \code{numeric vector} containing the theta values}

\item{process.desc}{A \code{character vector} containing the names of parameters.}

\item{freq}{A \code{double} indicating the frequency of the data.}
}
\description{
Convert from AR1 to GM and vice-versa
}
\author{
JJB
}
\keyword{internal}

