% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMWM.R
\name{gmwm-defunct}
\alias{compare.gmwm}
\alias{gmwm-defunct}
\title{Defunct function(s) in the gmwm package}
\usage{
compare.gmwm(...)
}
\arguments{
\item{...}{Grab old parameters}
}
\description{
These functions have been removed from the gmwm package.
}
\section{Details}{

\tabular{rl}{
  \code{compare.gmwm} \tab has been removed in favor of \code{\link{compare.models}}\cr
}
}

