% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GMWM_package.R
\docType{package}
\name{GMWM-package}
\alias{GMWM-package}
\title{GMWM Package}
\description{
GMWM is a Time Series Package
}
\details{
\tabular{ll}{
Package: \tab GMWM\cr
Type: \tab Package\cr
Version: \tab 0.6.7\cr
Date: \tab 2015-02-23\cr
License: \tab MIT + file LICENSE\cr
}
Some details should be added
}
\examples{
# Fill in at a later time
}
\author{
Stephane Guerrier \email{stephane@illinois.edu},
James Balamuta \email{balamut2@illinois.edu}

James Balamuta \email{balamut2@illinois.edu}
}
\seealso{
\code{avar}
}

