% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GMWM.R
\name{gmwm}
\alias{gmwm}
\title{GMWM for IMU, ARMA, SSM, and Robust}
\usage{
gmwm(model, data, model.type = "imu", fast = TRUE, augmented = FALSE,
  p = 0.025, robust = FALSE, eff = 0.6, B = 1000)
}
\arguments{
\item{model}{A \code{ts.model} object containing one of the allowed models. Not used if \code{type="arma"}}

\item{data}{A \code{matrix} or \code{data.frame} object with only column (e.g. \eqn{N \times 1}{ N x 1 })}

\item{model.type}{A \code{string} containing the type of GMWM needed e.g. IMU or SSM}

\item{fast}{A \code{boolean} indicating the type of covariance matrix solver. TRUE means bootstrap and FALSE means use a fast fourier transform (fft)}

\item{augmented}{A \code{boolean} indicating whether to add additional moments (e.g. mean for drift and variance for all other components).}

\item{p}{A \code{double} between 0 and 1 that correspondings to the \eqn{\frac{\alpha}{2}}{alpha/2} value for the wavelet confidence intervals.}

\item{robust}{A \code{boolean} indicating whether to use the robust computation (TRUE) or not (FALSE).}

\item{eff}{A \code{double} between 0 and 1 that indicates the efficiency.}

\item{B}{An \code{integer} to sample the space for IMU and SSM models to ensure optimal identitability.}
}
\value{
A \code{gmwm} object that contains:
\itemize{
 \item{}
 \item{}
 \item{}
}
}
\description{
GMM object
}
\details{
This function is under work. Some of the features are active. Others... Not so much.
What is NOT active:
1. Fast Computation of V
2. Augmented GMWM (additional moments)
}
\examples{
# AR
set.seed(1336)
n = 200
data = gen_ar1(n, phi=.99, sigma2 = 0.01) + gen_wn(n, sigma2=1)

# Models can contain specific parameters e.g.
adv.model = gmwm(AR1(phi = .99, sigma2 = 0.01) + WN(sigma2 = 0.01),
                            data)

# Or we can guess the parameters:
guided.model = gmwm(AR1() + WN(), data)

# Want to try different models?
guided.ar1 = gmwm(AR1(), data)

# Faster:
guided.ar1.wn.prev = update(guided.ar1, AR1()+WN())

# OR

# Create new GMWM object.
# Note this is SLOWER since the Covariance Matrix is recalculated.
guided.ar1.wn.new = gmwm(AR1()+WN(), data)

# ARMA case
set.seed(1336)
data = arima.sim(n = 200,
              list(ar = c(0.8897, -0.4858), ma = c(-0.2279, 0.2488)),
              sd = sqrt(0.1796))
guided.arma = gmwm(ARMA(2,2), data, model.type="ssm")
adv.arma = gmwm(ARMA(ar=(0.8897, -0.4858), ma = c(-0.2279, 0.2488), sigma2=0.1796),
                data, model.type="ssm")
}

