% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gmwm_arma_cpp}
\alias{gmwm_arma_cpp}
\title{GMWM ARMA}
\usage{
...(x)
}
\arguments{
\item{x}{A \code{vector} with dimensions N x 1.}

\item{type}{A \code{character string} indicating if the function should estimate an ARMA model ("ARMA"), a model for IMU sensor calibration ("IMU") or a state-space model ("SSM")}

\item{params}{A \code{vector} being numeric (if type = "ARMA") or character string (if type = "IMU" or type = "SSM")}

\item{robust}{A \code{bool} indicating if the function should provide a robust estimation of the model parameters (by default = FALSE).}
}
\value{
gmwm A \code{list} that contains:
\itemize{
 \item{par}{The estimated model parameters}
 \item{CI}{The 95\% confidence intervals for the estimated model parameters.}
}
}
\description{
This function uses the Generalized Method of Wavelet Moments to estimate the parameters of a time series model based on ARMA.
}
\details{
The function estimates a variety of time series models. If type = "ARMA" then the parameter vector (param) should
indicate the order of the AR process and of the MA process (i.e. param = c(AR,MA)).
If robust = TRUE the function takes the robust estimate of the wavelet variance to be used in the GMWM estimation procedure.
}
\examples{
...
}
\author{
JJB
}
\references{
Wavelet variance based estimation for composite stochastic processes, S. Guerrier and Robust Inference for Time Series Models: a Wavelet-Based Framework, S. Guerrier
}

