% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gmwm.R
\name{gmwm_guide}
\alias{gmwm}
\alias{gmwm_guide}
\title{GMWM for IMU, ARMA, SSM, and Robust}
\usage{
gmwm(desc, wvcov, signal, type = "imu", params = list(ar = 1, ma = 2),
  B = 1000)
}
\arguments{
\item{desc}{A \code{vec<string>} containing one of the allowed models. Not used if \code{type="arma"}}

\item{wvcov}{A \code{wvcov} object}

\item{signal}{The time series being studied}

\item{type}{A \code{string} containing the type of GMWM needed e.g. IMU, ARMA, or SSM}

\item{params}{A \code{list} containing the numbers of AR and MA parameters. Only used if \code{type="arma"}}

\item{B}{A \code{integer} to sample the space for IMU and SSM models to ensure AR1 identitability.}
}
\description{
GMM object
}
\examples{
# ARMA case
set.seed(1336)
n = 200
x = gen_ar1(n, phi=.1, sig2 = 1) + gen_ar1(n,phi=0.95, sig2 = .1)
decomp = modwt(x)
wv = wvar(decomp, robust = TRUE)
out = wvcov(decomp, wv, compute.v="diag")
save = gmwm(desc=c("AR1"), wvcov=out, signal=x, type="imu", B = 10000)

# ARMA case
set.seed(1336)
x=arima.sim(n = 200, list(ar = c(0.8897, -0.4858), ma = c(-0.2279, 0.2488)), sd = sqrt(0.1796))
decomp = modwt(x)
wv = wvar(decomp, robust = TRUE)
out = wvcov(decomp, wv, compute.v="diag")
save = gmwm(wvcov=out, signal=x, type="arma", params=list(ar=2,ma=2))
}

