% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gmwm_master_cpp}
\alias{gmwm_master_cpp}
\title{Master Wrapper for the GMWM Estimator}
\usage{
gmwm_master_cpp(data, theta, desc, objdesc, model_type, starting, alpha,
  compute_v, K, H, G, robust, eff, fullv)
}
\arguments{
\item{data}{A \code{vec} containing the data.}

\item{theta}{A \code{vec} with dimensions N x 1 that contains user-supplied initial values for parameters}

\item{desc}{A \code{vector<string>} indicating the models that should be considered.}

\item{objdesc}{A \code{field<vec>} containing a list of parameters (e.g. AR(1) = c(1,1), ARMA(p,q) = c(p,q,1))}

\item{model_type}{A \code{string} that represents the model transformation}

\item{starting}{A \code{bool} that indicates whether the supplied values are guessed (T) or are user-based (F).}

\item{alpha}{A \code{double} that handles the alpha level of the confidence interval (1-alpha)*100}

\item{compute_v}{A \code{string} that describes what kind of covariance matrix should be computed.}

\item{K}{An \code{int} that controls how many times theta is updated.}

\item{H}{An \code{int} that controls how many bootstrap replications are done.}

\item{G}{An \code{int} that controls how many guesses at different parameters are made.}

\item{robust}{A \code{bool} that indicates whether the estimation should be robust or not.}

\item{eff}{A \code{double} that specifies the amount of efficiency required by the robust estimator.}

\item{inference}{A \code{bool} that indicates whether inference should be run on the supplied model.}
}
\value{
A \code{field<mat>} that contains a list of ever-changing estimates...
}
\description{
This function generates WV, GMWM Estimator, and an initial test estimate.
}
\details{
The function estimates a variety of time series models. If type = "ARMA" then the parameter vector (param) should
indicate the order of the AR process and of the MA process (i.e. param = c(AR,MA)). If type = "IMU" or "SSM", then
parameter vector should indicate the characters of the models that compose the latent or state-space model. The model
options are:
\itemize{
  \item{"AR1"}{a first order autoregressive process with parameters \eqn{(\phi,\sigma^2)}{phi, sigma^2}}
  \item{"ARMA"}{an autoregressiveß moving average process with parameters \eqn{(\phi _p, \theta _q, \sigma^2)}{phi[p], theta[q], sigma^2}}
  \item{"DR"}{a drift with parameter \eqn{\omega}{omega}}
  \item{"QN"}{a quantization noise process with parameter \eqn{Q}}
  \item{"RW"}{a random walk process with parameter \eqn{\sigma^2}{sigma^2}}
  \item{"WN"}{a white noise process with parameter \eqn{\sigma^2}{sigma^2}}
}
If type = "ARMA", the function takes condition least squares as starting values; if type = "IMU" or type = "SSM" then
starting values pass through an initial bootstrap and pseudo-optimization before being passed to the GMWM optimization.
If robust = TRUE the function takes the robust estimate of the wavelet variance to be used in the GMWM estimation procedure.
}
\examples{
# Coming soon
}
\author{
JJB
}
\references{
Wavelet variance based estimation for composite stochastic processes, S. Guerrier and Robust Inference for Time Series Models: a Wavelet-Based Framework, S. Guerrier
}
\keyword{internal}

