% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/gmwm_logic.cpp, src/gmwm_logic.h
\name{gmwm_master_wv_cpp}
\alias{gmwm_master_wv_cpp}
\title{Master Wrapper for the GMWM Estimator (using WV and Omega as inputs)}
\usage{
gmwm_master_wv_cpp(
  wvar,
  N,
  expect_diff,
  omega,
  ranged,
  theta,
  desc,
  objdesc,
  model_type,
  starting,
  alpha,
  compute_v,
  K,
  H,
  G,
  robust,
  eff
)
}
\arguments{
\item{wvar}{A \code{mat} containing the empirical WV (col1), lower bound (col2) and upper bound (col3).}

\item{N}{A \code{integer} corresponding to the length of the original data.}

\item{expect_diff}{A \code{double} corresponding to the empirical mean of the first difference of the original data.}

\item{omega}{A \code{matrix} corresponding to the weight matrix of the GMWM estimator.}

\item{ranged}{A \code{double} corresponding to the scaled range of the original data (i.e. (max - min)/length).}

\item{theta}{A \code{vec} with dimensions N x 1 that contains user-supplied initial values for parameters}

\item{desc}{A \code{vector<string>} indicating the models that should be considered.}

\item{objdesc}{A \code{field<vec>} containing a list of parameters (e.g. AR(1) = c(1,1), ARMA(p,q) = c(p,q,1))}

\item{model_type}{A \code{string} that represents the model transformation}

\item{starting}{A \code{bool} that indicates whether the supplied values are guessed (T) or are user-based (F).}

\item{alpha}{A \code{double} that handles the alpha level of the confidence interval (1-alpha)*100}

\item{compute_v}{A \code{string} that describes what kind of covariance matrix should be computed.}

\item{K}{An \code{int} that controls how many times theta is updated.}

\item{H}{An \code{int} that controls how many bootstrap replications are done.}

\item{G}{An \code{int} that controls how many guesses at different parameters are made.}

\item{robust}{A \code{bool} that indicates whether the estimation should be robust or not.}

\item{eff}{A \code{double} that specifies the amount of efficiency required by the robust estimator.}
}
\value{
A \code{field<mat>} that contains a list of ever-changing estimates...
}
\description{
This function compute the GMWM Estimator, and an initial test estimate.
}
\references{
Wavelet variance based estimation for composite stochastic processes, S. Guerrier and Robust Inference for Time Series Models: a Wavelet-Based Framework, S. Guerrier
}
\author{
JJB, SG
}
\keyword{internal}
