% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gof_test}
\alias{gof_test}
\title{Compute the GOF Test}
\usage{
gof_test(theta, desc, objdesc, model_type, tau, v_hat, wv_empir)
}
\arguments{
\item{theta}{}

\item{psi}{}

\item{p}{}
}
\value{
A \code{vec} that has
\itemize{
\item Test Statistic
\item P-Value
\item DF
}
}
\description{
yaya
}

