% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gts.R
\name{gts}
\alias{gts}
\title{Create a GMWM TS Object}
\usage{
\method{gts}{}(data, name = "", freq = 1, unit = "sec")
}
\arguments{
\item{data}{A one column matrix, data.frame, or numeric vector.}

\item{name}{A string that provides an identifier to the data. If not supplied, the name of the first column will be used. Otherwise, it will be registered as "Dataset X"}

\item{freq}{A number that provides the rate of samples.}

\item{unit}{A string that contains the unit expression of the frequency.}
}
\value{
An S3 object with called gts with the following structure:
\itemize{
 \item{data}{Data set information}
 \item{name}{Name of the Dataset}
 \item{freq}{Numeric representation of frequency}
 \item{Unit}{String representation of the Unit}
}
}
\description{
Setups a time series oriented object that works well with graphing and summary utilities
}
\examples{
m = data.frame(rnorm(5))
gts(5)

x = gen.ts(WN(sigma2=1), 50)
x = gts(x)
}
\author{
JJB
}

