% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{head.imu}
\alias{head.gts}
\alias{head.imu}
\alias{head.lts}
\alias{tail.gts}
\alias{tail.imu}
\alias{tail.lts}
\title{Return the First or Last Part of a GMWM Object}
\usage{
\method{head}{imu}(x, n = 6L, ...)

\method{tail}{imu}(x, n = 6L, ...)

\method{head}{gts}(x, n = 6L, ...)

\method{tail}{gts}(x, n = 6L, ...)

\method{head}{lts}(x, n = 6L, ...)

\method{tail}{lts}(x, n = 6L, ...)
}
\arguments{
\item{x}{A \code{gts}, \code{imu}, or \code{lts} object.}

\item{n}{A \code{integer} indicating how many observations should be displayed}

\item{...}{Arguments to be passed to or from other methods.}
}
\value{
The method will return a matrix with \eqn{n} observations.
}
\description{
Returns the first or last \eqn{n} observations of a \code{gts}, \code{imu}, or \code{lts} object.
}
\author{
JJB
}

