% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/imu.R
\docType{data}
\name{imu}
\alias{imu}
\title{IMU Data}
\format{A \code{data.frame} with 873,684 observations and 6 variables. The variables are as follows:
\itemize{
 \item{Gyro. X}{Gyroscope X-Axis}
 \item{Gyro. Y}{Gyroscope Y-Axis}
 \item{Gyro. Z}{Gyroscope Z-Axis}
 \item{Accel. X}{Accelerometer X-Axis}
 \item{Accel. Y}{Accelerometer Y-Axis}
 \item{Accel. Z}{Accelerometer Z-Axis}
}}
\source{
Where's the data frame?
}
\usage{
imu
}
\description{
This data set contains gyroscope data
}
\examples{
\dontrun{
data(imu)
}
}
\author{
JJB
}

