% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/imu_graphs.r
\name{imu2WV}
\alias{imu2WV}
\title{Obtain the Wavelet Variance for imu data}
\usage{
imu2WV(object)
}
\arguments{
\item{object}{A \code{data.frame} or \code{matrix} that contains 6 columns.}
}
\value{
A \code{data.frame} that contains the formatted data.
}
\description{
Generates the Wavelet Variance for each column in the data set.
}
\examples{
\dontrun{
data(imu)
df = imu2WV(imu)
}
}

