% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imu.R
\name{imu_}
\alias{imu_}
\title{Internal IMU Object Construction}
\usage{
imu_(object, ngyros, nacces, axis)
}
\arguments{
\item{object}{A \code{matrix} with dimensions N x length(index)}

\item{ngyros}{A \code{integer} containing the number of gyroscopes}

\item{axis}{A \code{vector} unique representation of elements e.g. x,y,z or x,y or x.}

\item{naccess}{A \code{integer} containing the number of accelerometers}
}
\value{
An \code{imu} object class.
}
\description{
Internal quick build for imu object.
}
\keyword{internal}

