% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{imu_time}
\alias{imu_time}
\title{Pulls the IMU time from the IMU object}
\usage{
imu_time(x)
}
\arguments{
\item{x}{A \code{imu} object}
}
\value{
A \code{vector} with numeric information.
}
\description{
Helper function for the IMU object to access \code{rownames()} with a numeric conversion.
}
