% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inference_summary}
\alias{inference_summary}
\title{Compute the Inference Summary}
\usage{
inference_summary(theta, desc, objdesc, model_type, tau, D, v_hat, omega,
  wv_empir, N, alpha, ci_bootstrap, B, robust, eff)
}
\value{
A \code{field<mat>} where
\itemize{
\item CI
\itemize{
\item Lower Bound
\item Upper Bound
\item Standard Error
}
\item GOF
\itemize{
\item Test Statistic
\item P-Value
\item DF
}
}
}
\description{
Calls CI and GOF Test computation functions.
}

