% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_install.R
\name{install_}
\alias{install_}
\title{Light weight package install}
\usage{
install_(user, pkg.name)
}
\arguments{
\item{user}{A \code{string} containing the user name}

\item{pkg.name}{A \code{string} that contains the repository name}
}
\description{
Downloads an R package from github via the zip archive option,
unzips the repository, builds the R package source, and 
install the package
}
\details{
This is a lightweight standalone method to download and install packages.
It is much safer to use devtools, though, it is one additional "download".
}
\examples{
\dontrun{
# Online install
install_("SMAC-Group","imudata")
}
}
\author{
JJB
}

