% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_install.R
\name{install_datapkg}
\alias{install_datapkg}
\title{Install Data Package}
\usage{
install_datapkg(type = "ONL", loc = NULL)
}
\arguments{
\item{type}{A \code{string} with value \code{"ONL"} or \code{"LOCAL"}}

\item{loc}{A \code{string} that contains the file location.}
}
\description{
Downloads and Install the Data Package to use with the GMWM Package
}
\examples{
\dontrun{
# Online install
install_datapkg()

# Local install
install_datapkg("LOCAL","C:/Users/James/Documents/datapkg_x.y.z.tar.gz")
}
}
\author{
JJB
}

