% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_install.R
\name{install_imudata}
\alias{install_imudata}
\title{Install IMU Data Package}
\usage{
install_imudata(type = "ONL", loc = NULL)
}
\arguments{
\item{type}{A \code{string} with value \code{"ONL"} or \code{"LOCAL"}}

\item{loc}{A \code{string} that contains the file location.}
}
\description{
Downloads and Install the IMU Data Package to use with the GMWM Package
}
\details{
To obtain the IMU Package
}

