% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is.rstudio}
\alias{is.rstudio}
\title{Is R Open in RStudio?}
\usage{
is.rstudio()
}
\value{
A \code{logical} value that indicates whether R is open in RStudio.
}
\description{
Detects whether R is open in RStudio.
}
\examples{
is.rstudio()
}
\author{
JJB
}
\keyword{internal}

