% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is.validCompEffObj}
\alias{is.validCompEffObj}
\title{Validity of the Object List for \code{compare.eff}}
\usage{
is.validCompEffObj(obj.list)
}
\arguments{
\item{obj.list}{A \code{list} of \code{gmwm} object}
}
\description{
Check whether the object list is valid for function \code{compare.eff}.
}
\details{
A valid object list should contain \code{gmwm} objects constructed by the same data
and the same model.
}
\keyword{internal}

