% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Check Integer Values}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{A `numeric` value which needs to be checked to understand if it is an `integer`.}
}
\value{
A `boolean` value that specifies whether the supplied value is an `integer` or not.
}
\description{
This function is taken from the examples of the `integer` function from the `base` package.
}
\examples{
is.wholenumber(4.2)
is.wholenumber(6)
is.wholenumber(seq(1, 5, by = 0.5))
}
\author{
`base` R package
}
