% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is.gts}
\alias{is.gmwm}
\alias{is.gts}
\alias{is.imu}
\alias{is.lts}
\alias{is.ts.model}
\alias{is.wvar}
\title{Is GMWM Object}
\usage{
is.gts(x)

is.imu(x)

is.lts(x)

is.gmwm(x)

is.wvar(x)

is.ts.model(x)
}
\arguments{
\item{x}{A \code{gts}, \code{imu}, \code{lts} object.}
}
\value{
A \code{logical} value that indicates whether the object is of that class (TRUE) or not (FALSE).
}
\description{
Is the object a
\code{gts}, \code{imu}, \code{lts}, \code{wvar}, or \code{gmwm} object?
}
\details{
Uses \code{\link[base]{inherits}} over \code{\link[methods]{is}} for speed.
}
\author{
JJB
}

