% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/analytical_matrix_derivatives.cpp, src/analytical_matrix_derivatives.h
\name{jacobian_arma}
\alias{jacobian_arma}
\title{Calculates the Jacobian for the ARMA process}
\usage{
jacobian_arma(theta, p, q, tau)
}
\arguments{
\item{theta}{A \code{vec} that contains all the parameter estimates.}

\item{p}{A \code{int} that indicates the number of AR coefficients}

\item{q}{A \code{int} that indicates the number of MA coefficients.}

\item{tau}{A \code{vec} that lists the scales of the process e.g. 2^(1:J)}
}
\value{
A \code{mat} that returns the numerical jacobian of the ARMA process.
}
\description{
Figure out the Jacobian for an ARMA process.
}
\keyword{internal}

