% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lm_dr}
\alias{lm_dr}
\title{Linear Regression with Drift}
\usage{
lm_dr(x)
}
\arguments{
\item{x}{A \code{vec} of length \eqn{N\times 1}{N x 1} containing the responses.}
}
\value{
A \code{field<vec>} with:
\describe{
  \item{coef}{Coefficients}
  \item{resid}{Residuals}
  \item{sigma2}{Sigma^2}
}
}
\description{
Perform a linear regression with drift.
}
\examples{
x = 1:10
y = cumsum(rep(.7,10))

lm_dr(y)[[1]]

coef(lm(y~x-1))

}
\keyword{internal}
