% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logit}
\alias{logit}
\title{Logit Function}
\usage{
logit(x)
}
\arguments{
\item{x}{A \code{vec} containing probabilities (e.g. -1 <= x <= 1)}
}
\value{
A \code{vec} containing logit terms.
}
\description{
This function computes the logit link function.
}
\examples{
x.sim = runif(100)
logit(x.sim)
}
\keyword{internal}

