% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logit2_inv}
\alias{logit2_inv}
\title{Logit2 Inverse Function}
\usage{
logit2_inv(x)
}
\arguments{
\item{x}{A \code{vec} containing real numbers.}
}
\value{
A \code{vec} containing logit probabilities.
}
\description{
This function computes the inverse of a logit transformation of the parameters.
}
\examples{
x.sim = rnorm(100)
logit_inv(x.sim)
}
\keyword{internal}

