% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{m2_drift}
\alias{m2_drift}
\title{Second moment DR}
\usage{
m2_drift(omega, n_ts)
}
\arguments{
\item{omega}{A \code{double} corresponding to variance of drift.}

\item{n_ts}{An \code{int} indicating the length of the time series.}
}
\value{
A \code{vec} containing the second moment of the drift.
}
\description{
This function computes the second moment of a drift process.
}
\examples{
m2_drift(1, 200)
}
\keyword{internal}
