% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.model.R
\name{MA1}
\alias{MA1}
\title{Create an Moving Average 1 [MA(1)] Process}
\usage{
MA1(theta = NULL, sigma2 = 1)
}
\arguments{
\item{theta}{A \code{double} value for the \eqn{\phi}{phi} of an MA1 process.}

\item{sigma2}{A \code{double} value for the variance, \eqn{\sigma ^2}{sigma^2}, of a WN process.}
}
\value{
An S3 object with called ts.model with the following structure:
\describe{
 \item{process.desc}{Used in summary: "MA1","SIGMA2"}
 \item{theta}{\eqn{\theta}{theta}, \eqn{\sigma^2}{sigma^2}}
 \item{plength}{Number of Parameters}
 \item{desc}{"MA1"}
 \item{obj.desc}{Depth of Parameters e.g. list(1,1)}
 \item{starting}{Guess Starting values? TRUE or FALSE (e.g. specified value)}
}
}
\description{
Setups the necessary backend for the MA1 process.
}
\examples{
MA1()
MA1(theta=.32, sigma2=1.3)
}
\author{
JJB
}
