% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ma1_to_wv}
\alias{ma1_to_wv}
\title{Moving Average Order 1 (MA(1)) to WV}
\usage{
ma1_to_wv(theta, sig2, tau)
}
\arguments{
\item{theta}{A \code{double} corresponding to the moving average term.}

\item{sig2}{A \code{double} the variance of the process.}

\item{tau}{A \code{vec} containing the scales e.g. 2^tau}
}
\value{
A \code{vec} containing the wavelet variance of the MA(1) process.
}
\description{
This function computes the WV (haar) of a Moving Average order 1 (MA1) process.
}
\examples{
ntau = 7
tau = 2^(1:ntau)
wv.theo = ma1_to_wv(1, tau)
}

