% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mean_diff}
\alias{mean_diff}
\title{Mean of the First Difference of the Data}
\usage{
mean_diff(x)
}
\arguments{
\item{x}{A \code{vec} containing the data}
}
\value{
A \code{double} that contains the mean of the first difference of the data.
}
\description{
The mean of the first difference of the data
}
\examples{
x=rnorm(10)
mean_diff(x)
}
\keyword{internal}
