% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/ts_model_cpp.cpp, src/ts_model_cpp.h
\name{model_process_desc}
\alias{model_process_desc}
\title{Generate the ts model object's process desc}
\usage{
model_process_desc(desc)
}
\arguments{
\item{desc}{A \code{vector<string>} that contains a list of the strings of each process.}
}
\value{
A \code{vector<string>} with a list of descriptive values to label the estimate matrix with
}
\description{
Creates the ts.model's process desc
}
\details{
This function currently does NOT support ARMA(P,Q) models. 
That is, there is no support for ARMA(P,Q), AR(P), or MA(Q).
There is support for ARMA11, AR1, MA1, GM, WN, DR, QN, and RW.
}
\keyword{internal}
