% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{model_score}
\alias{model_score}
\title{Model Score}
\usage{
model_score(A, D, omega, v_hat, diff)
}
\arguments{
\item{A}{A \code{mat} that contains the first derivatives of the processes}

\item{omega}{A \code{mat} that contains the omega used when calculating the GMWM}

\item{v_hat}{A \code{mat} that contains the covariance matrix}

\item{diff}{A \code{vec} that is the difference of the WV empirical and WV theoretical}

\item{At_j}{A \code{mat} that contains the second derivative of each process}
}
\value{
A \code{vec}
}
\description{
Calculates the modeling score of a GMWM
}
\details{
The equation is slightly different than that stated in the paper due to the bootstrap already incorporating in
N.
}

