% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Multiple plot function}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{ggplot objects can be passed in}

\item{plotlist}{A \code{list} of ggplot objects.}

\item{cols}{An \code{integer} containing the number of columns in layout}

\item{layout}{A \code{matrix} specifying the layout. If present, 'cols' is ignored.}
}
\value{
Creates a combination ggplot panel.
}
\description{
Enables multiple ggplot objects to appear in a panel.
}
\details{
If the layout is something like matrix(c(1,2,3,3), nrow=2, byrow=TRUE),
then plot 1 will go in the upper left, 2 will go in the upper right, and
3 will go all the way across the bottom.

See the R Graphics Cookbook for more information
}
\author{
Winston Chang
}
\keyword{internal}

