% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ncol.gts}
\alias{NCOL.gts}
\alias{NCOL.imu}
\alias{NCOL.lts}
\alias{NROW.gts}
\alias{NROW.imu}
\alias{NROW.lts}
\alias{ncol.gts}
\alias{ncol.imu}
\alias{ncol.lts}
\alias{nrow.gts}
\alias{nrow.imu}
\alias{nrow.lts}
\title{The Number of Rows/Columns of a GMWM Object}
\usage{
ncol.gts(x)

NCOL.gts(x)

nrow.gts(x)

NROW.gts(x)

ncol.imu(x)

NCOL.imu(x)

nrow.imu(x)

NROW.imu(x)

ncol.lts(x)

NCOL.lts(x)

nrow.lts(x)

NROW.lts(x)
}
\arguments{
\item{x}{A \code{gts}, \code{imu}, \code{lts} object.}
}
\value{
An \code{\link[base]{integer}} of length 1 or \code{\link[base]{NULL}}
}
\description{
Returns either the number of rows or columns of a 
\code{gts}, \code{imu}, or \code{lts} object.
}
\author{
JJB
}

