% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{obj_extract}
\alias{obj_extract}
\title{Extract Object}
\usage{
obj_extract(theta, objdesc, cur_position)
}
\arguments{
\item{theta}{A \code{vec} containing the theta values.}

\item{objdesc}{A \code{vec} at the desc point.}

\item{cur_position}{An \code{integer} at the current position.}
}
\value{
A \code{field<vec>} containing the breakdown of the object.
}
\description{
Extracts the object information and returns it.
}
\keyword{internal}
