% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\docType{methods}
\name{objFun_find_biwc}
\alias{objFun_find_biwc}
\title{Objective Function for Tuning Constant}
\usage{
objFun_find_biwc(crob, eff)
}
\arguments{
\item{crob}{A \code{double} that indicates the value to optimize. Normally between [0,5].}

\item{eff}{A \code{double} that represents the efficiency.}
}
\value{
A \code{double} that contains the biwc.
}
\description{
Objective function that finds tuning constant
}
\details{
This function is not available from within R.
}

