% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{opt_n_gof_bootstrapper}
\alias{opt_n_gof_bootstrapper}
\title{Bootstrap for Optimism and GoF}
\usage{
opt_n_gof_bootstrapper(theta, desc, objdesc, scales, model_type, N, robust, eff,
  alpha, H)
}
\arguments{
\item{theta}{A \code{vector} with dimensions N x 1 that contains user-supplied initial values for parameters}

\item{desc}{A \code{vector<string>} indicating the models that should be considered.}

\item{objdesc}{A \code{field<vec>} that contains an object description (e.g. values) of the model.}

\item{scales}{A \code{vec} containing the scales of the process.}

\item{model_type}{A \code{string} containing the model type either: SSM or IMU}

\item{N}{A \code{int} indicating how long the integer is.}

\item{robust}{A \code{bool} indicating robust (T) or classical (F).}

\item{eff}{A \code{double} that handles efficiency.}

\item{H}{A \code{int} that indicates how many bootstraps should be obtained.}
}
\value{
A \code{vec} that contains the parameter estimates from GMWM estimator.
}
\description{
Using the bootstrap approach, we simulate a model based on user supplied parameters, obtain the wavelet variance, and then V.
}
\details{
Expand in detail...
}
\examples{
# Coming soon
}
\author{
JJB
}
\keyword{internal}
