% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{order_AR1s}
\alias{order_AR1s}
\title{Order AR1s by size of phi.}
\usage{
order_AR1s(theta, desc, objdesc)
}
\value{
A \code{vec} that has AR1s shown in descending parameter value.
}
\description{
Changes the order of AR1s in a string by size.
}

