% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{packageVersionCRAN}
\alias{packageVersionCRAN}
\title{Latest Version of Package on CRAN}
\usage{
packageVersionCRAN(pkg, cran_url = "http://cran.r-project.org/web/packages/")
}
\arguments{
\item{pkg}{Name of Package}

\item{cran_url}{URL to CRAN Packages}
}
\value{
A \code{vector} of \code{string}s that contain:
\itemize{
\item the verison of the package on cran
\item release date of the package on cran
}
}
\description{
Determines the version number on cran via obtaining the packages page
}
\examples{
library(gmwm)
packageVersion("gmwm")
packageVersionCRAN("gmwm")

}
\author{
JJB
}
\keyword{internal}
