% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{paperSetting}
\alias{paperSetting}
\title{Frequent Graph Setting for Paper}
\usage{
paperSetting()
}
\value{
A ggplot2 panel containing the frequent graph setting for paper.
}
\description{
This function sets some parameters such as plot.margin.
}
\keyword{internal}

