% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{placeLegend}
\alias{placeLegend}
\title{Place Legend}
\usage{
placeLegend(wv_1, low_n, high_n)
}
\arguments{
\item{wv_1}{A \code{double} that indicates the first value of \code{wv.empir}}

\item{low_n}{A \code{doble} that indicates the last value of \code{ci.low}}

\item{high_n}{A \code{dobule} that indicates the last value of \code{ci.high}}
}
\value{
A numeric vector containing 4 elements. The first two elements indicate legend.justification, the last two elements indicate legend.position (see \code{?theme}).
}
\description{
This function decides where the legend should be put (top left or bottom left)
}
\keyword{internal}
