% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf.R
\name{plot.ACF}
\alias{autoplot.ACF}
\alias{plot.ACF}
\title{Auto-Covariance and Correlation Functions}
\usage{
\method{plot}{ACF}(x, ci = 0.95, show.ci = TRUE, ...)

\method{autoplot}{ACF}(object, ci = 0.95, show.ci = TRUE, ...)
}
\arguments{
\item{x, object}{An \code{"ACF"} object from \code{\link{ACF}}.}

\item{ci}{A \code{double} containing the 1-alpha level. Default is 95%}

\item{show.ci}{A \code{bool} indicating whether to show confidence region}

\item{...}{Additional parameters}
}
\value{
An \code{array} of dimensions \eqn{N \times S \times S}{N x S x S}.
}
\description{
The acf function computes the estimated
autocovariance or autocorrelation for both univariate and multivariate cases.
}

