% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avar.r
\name{plot.avar}
\alias{plot.avar}
\title{Plot Allan Variance}
\usage{
\method{plot}{avar}(x, ...)
}
\arguments{
\item{x}{A \code{avar} object.}

\item{...}{Arguments to be passed to methods}
}
\value{
ggplot2 graph
}
\description{
Displays a plot containing the allan variance
}
\examples{
set.seed(999)
x=rnorm(100)
out = avar(x)
plot( out )
}
\author{
JJB
}

