% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadam.R
\name{plot.hadam}
\alias{plot.hadam}
\title{Plot Hadamard Variance}
\usage{
\method{plot}{hadam}(x, ...)
}
\arguments{
\item{x}{A \code{hadam} object.}

\item{...}{Arguments to be passed to methods}
}
\value{
ggplot2 graph
}
\description{
Displays a plot containing the Hadamard variance
}
\examples{
set.seed(999)
x=rnorm(100)
out = hadam(x)
plot( out )
}
\author{
JJB
}

