% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/imu_graphs.r
\name{plot.imu}
\alias{plot.imu}
\title{Plot the Wavelet Variances in separate and combined way}
\usage{
\method{plot}{imu}(x, separate = TRUE, CI = TRUE, transparence = 0.1,
  color.line = "black", color.CI = "#003C7D", line.type = "solid",
  graph.title = "Haar Wavelet Variance Representation",
  graph.title.size = 15, axis.label.size = 13, axis.tick.size = 11,
  title.x.axis = expression(paste("Scale ", tau)),
  title.y.axis = expression(paste("Wavelet Variance ", nu)),
  legend.title = "Axis", legend.key.size = 1.3, legend.title.size = 13,
  legend.text.size = 13, ...)
}
\arguments{
\item{x}{An \code{imu} object (use \code{imu2WV()} to create it)}

\item{separate}{A \code{boolean} that indicates whether the graphs should be split or combined.}

\item{CI}{A \code{boolean} that indicates whether the confidence interval should be plotted.}

\item{transparence}{A \code{double} that ranges from 0 to 1 that controls the transparency of the graph}

\item{color.line}{A \code{string} that indicates the color of the line drawn (only work when graphs are seperate)}

\item{color.CI}{A \code{string} that indicates the color of the confidence interval (e.g. black, red, #003C7D, etc.) (only work when graphs are seperate)}

\item{line.type}{A \code{string} that indicates the type of line (e.g. solid, dotted, etc.)}

\item{graph.title}{A \code{string} that indicates the title of the graph}

\item{graph.title.size}{An \code{integer} that indicates the size of title.}

\item{axis.label.size}{An \code{integer} that indicates the size of label}

\item{axis.tick.size}{An \code{integer} that indicates the size of tick mark}

\item{title.x.axis}{A \code{string} that indicates the label on x axis}

\item{title.y.axis}{A \code{string} that indicates the label on y axis}

\item{legend.title}{A \code{string} that indicates the title of legend (only work when graphs are combined)}

\item{legend.key.size}{A \code{double} that indicates the size of key (in centermeters) on legend (only work when graphs are combined)}

\item{legend.title.size}{An \code{integer} that indicates the size of title on legend (only work when graphs are combined)}

\item{legend.text.size}{An \code{integer} that indicates the size of key label on legend (only work when graphs are combined)}

\item{...}{Additional options}
}
\value{
A panel containing the graph of an IMU sensor.
}
\description{
Creates a graph of the wavelet variance object. The graphs can be either split (features a CI) or combined (all gyroscopes together and all accelerometer).
}
\examples{
\dontrun{
data(imu)
df = imu2WV(imu)
plot(df)
plot(df, separate=FALSE)
}
}

