% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection_methods.R
\name{plot.rank.models}
\alias{plot.rank.models}
\title{Graph `rank.models` object}
\usage{
\method{plot}{rank.models}(
  x,
  process.decomp = FALSE,
  background = "white",
  CI = T,
  transparence = 0.1,
  bw = F,
  CI.color = "#003C7D",
  line.type = NULL,
  line.color = NULL,
  point.size = NULL,
  point.shape = NULL,
  title = NULL,
  title.size = 15,
  axis.label.size = 13,
  axis.tick.size = 11,
  axis.x.label = expression(paste("Scale ", tau)),
  axis.y.label = expression(paste("Wavelet Variance ", nu)),
  legend.title = "",
  legend.label = NULL,
  legend.key.size = 1,
  legend.title.size = 13,
  legend.text.size = 13,
  ...
)
}
\description{
Creates a `rank.models` graph.
}
