% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wvar.R
\name{plot.wvar}
\alias{plot.wvar}
\title{Wrapper to ggplot Wavelet Variances Graph}
\usage{
\method{plot}{wvar}(x, ...)
}
\arguments{
\item{x}{A \code{wvar} object.}

\item{...}{other arguments passed to specific methods}
}
\value{
A ggplot2 graph containing the wavelet variances.
}
\description{
Creates the wavelet variance graph
}
\examples{
set.seed(999)
x=rnorm(100)
out = wvar(modwt(x))
plot( out )
}
\author{
JJB
}
\seealso{
\code{\link{autoplot.wvar}}
}

