% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ts.model.R
\name{+.ts.model}
\alias{+.ts.model}
\title{Add ts.model objects together}
\usage{
\method{+}{ts.model}(x, y)
}
\arguments{
\item{x}{A \code{ts.model} object}

\item{y}{A \code{ts.model} object}
}
\value{
An S3 object with called ts.model with the following structure:
\itemize{
 \item{desc}{combined x, y desc}
 \item{theta}{combined x, y theta}
 \item{plength}{Number of Parameters}
 \item{obj}{Depth of Parameters e.g. list(1, c(1,1), c(length(ar),length(ma),1) )}
 \item{adv}{Advanced Input TRUE or FALSE (e.g. specified value)}
}
}
\description{
Sets up the necessary backend for combining ts.model objects.
}
\examples{
DR()+WN()
AR1(phi=.3,sigma=.2)
}
\author{
JJB
}

