% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GMWM.R
\name{predict.gmwm}
\alias{predict.gmwm}
\title{Predict future points in the time series using the solution of the Generalized Method of Wavelet Moments}
\usage{
\method{predict}{gmwm}(object, n.ahead, ...)
}
\arguments{
\item{object}{GMWM}

\item{n.ahead}{Number of observations to guess.}

\item{data}{Set used to create the GMWM.}
}
\description{
Creates a prediction using the estimated values of GMWM through the ARIMA function within R.
}

