% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_selection_methods.R
\name{pretty.model.score}
\alias{pretty.model.score}
\title{Formats the model score matrix}
\usage{
\method{pretty}{model.score}(out, model.names)
}
\arguments{
\item{out}{A \code{list} containing the model matrix}

\item{num.models}{A \code{integer} that indicates the number of models fit.}
}
\value{
An updated matrix in the ith position of the list.
}
\description{
The model score matrix receives the appropriate model numbering, col descriptors, and ordering of models.
}

