% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection_methods.R
\name{print.auto.imu}
\alias{print.auto.imu}
\title{Print function for auto.imu object}
\usage{
\method{print}{auto.imu}(x, ...)
}
\arguments{
\item{x}{A \code{auto.imu} object.}

\item{...}{Additional parameters}
}
\description{
Prints the auto.imu function nicely.
}
\keyword{internal}

