% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avar.r
\name{print.avar}
\alias{print.avar}
\title{Prints Allan Variance}
\usage{
\method{print}{avar}(x, ...)
}
\arguments{
\item{x}{A \code{avar} object.}

\item{...}{Arguments to be passed to methods}
}
\value{
console output
}
\description{
Displays the allan variance information
}
\examples{
set.seed(999)
x=rnorm(100)
out = avar(x)
print( out )
}
\author{
JJB
}

