% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMWM.R
\name{print.gmwm}
\alias{print.gmwm}
\title{Print gmwm object}
\usage{
\method{print}{gmwm}(x, ...)
}
\arguments{
\item{x}{A \code{GMWM} object}

\item{...}{Other arguments passed to specific methods}
}
\value{
Text output via print
}
\description{
Displays information about GMWM object
}
\examples{
\dontrun{
# AR
set.seed(1336)
n = 200
xt = gen.gts(AR1(phi=.1, sigma2 = 1) + AR1(phi=0.95, sigma2 = .1),n)
mod = gmwm(AR1()+AR1(), data=xt, model.type="imu")
print(mod)
}
}
\author{
JJB
}
\keyword{internal}

