% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadam.R
\name{print.hadam}
\alias{print.hadam}
\title{Prints Hadamard Variance}
\usage{
\method{print}{hadam}(x, ...)
}
\arguments{
\item{x}{A \code{hadam} object.}

\item{...}{Arguments to be passed to methods}
}
\value{
console output
}
\description{
Displays the all Hadamard variance information
}
\examples{
set.seed(999)
x = rnorm(100)
out = hadam(x)
print( out )
}
\author{
JJB
}

