% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection_methods.R
\name{print.rank.models}
\alias{print.rank.models}
\title{Print function for rank.models object}
\usage{
\method{print}{rank.models}(x, ...)
}
\arguments{
\item{x}{A \code{rank.models} object.}

\item{...}{Additional parameters}
}
\description{
Prints the rank.models function nicely.
}
\keyword{internal}

