% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ts.model.R
\name{print.ts.model}
\alias{print.ts.model}
\title{Multiple a ts.model by constant}
\usage{
\method{print}{ts.model}(x, ...)
}
\arguments{
\item{x}{A \code{numeric} value}

\item{...}{further arguments passed to or from other methods.}
}
\value{
An S3 object with called ts.model with the following structure:
\itemize{
 \item{desc}
 \item{theta}
}
}
\description{
Sets up the necessary backend for creating multiple model objects.
}
\examples{
DR()+WN()+RW()+AR1()+ARMA(1,2)
}
\author{
JJB
}

