% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wvar.R
\name{print.wvar}
\alias{print.wvar}
\title{Print Wavelet Variances}
\usage{
\method{print}{wvar}(object, ...)
}
\arguments{
\item{object}{A \code{wvar} object.}

\item{...}{Not used...}
}
\value{
Summary table
}
\description{
Displays the summary table of wavelet variance.
}
\examples{
set.seed(999)
x=rnorm(100)
out = wvar(modwt(x))
print( out )
}
\author{
JJB
}

