% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compute.cov.R
\name{print.wvcov}
\alias{print.wvcov}
\title{Print Asymptotic Covariance Matrix}
\usage{
print.wvcov(object, ...)
}
\arguments{
\item{object}{A \code{wvcov} object}

\item{...}{Not used}
}
\value{
Prints the modwt matrix decomposition
}
\description{
Places the Asymptotic Covariance Matrix in print form.
}
\examples{
set.seed(999)
x=rnorm(100)
decomp = modwt(x)
wv = wvar(decomp)
print(wvcov(decomp,wv,compute.v="diag"))
}
\author{
JJB
}

