% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{print.imu}
\alias{outf}
\alias{print.gts}
\alias{print.imu}
\alias{print.lts}
\title{Print GMWM Data Object}
\usage{
\method{print}{imu}(x, obs = 10L, row.names = TRUE, ...)

\method{print}{lts}(x, obs = 10L, row.names = TRUE, ...)

\method{print}{gts}(x, obs = 10L, row.names = TRUE, ...)

outf(x, obs = 10L, row.names = TRUE)
}
\arguments{
\item{x}{A \code{gts}, \code{imu}, \code{lts} object.}

\item{obs}{A \code{integer} the specifies how many from the beginning and end of the data set to show.}

\item{row.names}{A \code{boolean} that indicates whether row names should be displayed or surpressed.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A \code{logical} value that indicates whether the object is of that class (TRUE) or not (FALSE).
}
\description{
Pretty formatting for \code{gts}, \code{imu}, and \code{lts} objects.
}
\author{
JJB
}

