% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pseudo_logit}
\alias{pseudo_logit}
\title{Pseudo Logit Function}
\usage{
pseudo_logit(x)
}
\arguments{
\item{x}{A \code{vec} containing probabilities (e.g. 0 <= x <= 1)}
}
\value{
A \code{vec} containing logit terms.
}
\description{
This function compute the link function to constrain parameters to a positive domain.
}
\examples{
x.sim = runif(100)
pseudo_logit(x.sim)
}
\keyword{internal}

