% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pseudo_logit_inv}
\alias{pseudo_logit_inv}
\title{Pseudo Logit Inverse Function}
\usage{
pseudo_logit_inv(x)
}
\arguments{
\item{x}{A \code{vec} containing real numbers.}
}
\value{
A \code{vec} containing logit probabilities.
}
\description{
This function computes the pseudo inverse of a logit transformation of the parameters in order to constrain them to a positive domain
}
\examples{
x.sim = rnorm(100)
pseudo_logit_inv(x.sim)
}
\keyword{internal}

