% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{returnRobustEff}
\alias{returnRobustEff}
\title{Get the Computation Method and Efficiency of \code{gmwm} Object}
\usage{
returnRobustEff(x)
}
\arguments{
\item{x}{A \code{gmwm} object}
}
\description{
The computation method (classical/robust) and efficiency will be returned in a certain format.
}
\details{
Used in \code{compare.eff()}.
}
\examples{
n = 1000
x = gen.gts(AR1(phi = .1, sigma2 = 1) + AR1(phi = 0.95, sigma2 = .1), n)
GMWM1 = gmwm(2*AR1()+RW(), data = x, robust  = TRUE, eff = 0.9)
GMWM2 = gmwm(2*AR1()+RW(), data = x, robust  = FALSE)

returnRobustEff(GMWM1)
returnRobustEff(GMWM2)
}
\keyword{internal}

