% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rw_to_wv}
\alias{rw_to_wv}
\title{Random Walk to WV}
\usage{
rw_to_wv(sig2, Tau)
}
\arguments{
\item{sig2}{A \code{double} corresponding to variance of RW}

\item{Tau}{A \code{vec} containing the scales e.g. 2^tau}
}
\value{
A \code{vec} containing the wavelet variance of the random walk.
}
\description{
This function compute the WV (haar) of a Random Walk process
}
\examples{
x.sim = cumsum(rnorm(100000))
waveletVariance(x.sim)
wv.theo = rw_to_wv(1,x.sim)
lines(wv.theo$scales,wv.theo$WV, col = "red")
}

