% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sarma_calculate_spadding}
\alias{sarma_calculate_spadding}
\title{Calculates Length of Seasonal Padding}
\usage{
sarma_calculate_spadding(np, nq, nsp, nsq, ns)
}
\arguments{
\item{np}{An \code{unsigned int} containing the number of non-seasonal phi parameters.}

\item{nq}{An \code{unsigned int} containing the number of non-seasonal theta parameters.}

\item{nsp}{An \code{unsigned int} containing the number of seasonal phi parameters.}

\item{nsq}{An \code{unsigned int} containing the number of seasonal theta parameters.}
}
\value{
A \code{vec} with rows:
\describe{
 \item{p}{Number of phi parameters}
 \item{q}{Number of theta parameters}
}
}
\description{
Computes the total phi and total theta vector length.
}
\seealso{
\code{\link{sarma_components}}
}
\keyword{internal}
