% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sarma_objdesc}
\alias{sarma_objdesc}
\title{Create the ts.model obj.desc given split values}
\usage{
sarma_objdesc(ar, ma, sar, sma, s, i, si)
}
\arguments{
\item{ar}{A \code{vec} containing the non-seasonal phi parameters.}

\item{ma}{A \code{vec} containing the non-seasonal theta parameters.}

\item{sar}{A \code{vec} containing the seasonal phi parameters.}

\item{sma}{A \code{vec} containing the seasonal theta parameters.}

\item{s}{An \code{unsigned integer} containing the frequency of seasonality.}

\item{i}{An \code{unsigned integer} containing the number of non-seasonal differences.}

\item{si}{An \code{unsigned integer} containing the number of seasonal differences.}
}
\value{
A \code{vec} with rows:
\describe{
\item{np}{Number of Non-Seasonal AR Terms}
\item{nq}{Number of Non-Seasonal MA Terms}
\item{nsp}{Number of Seasonal AR Terms}
\item{nsq}{Number of Seasonal MA Terms}
\item{nsigma}{Number of Variances (always 1)}
\item{s}{Season Value}
\item{i}{Number of non-seasonal differences}
\item{si}{Number of Seasonal Differences}
}
}
\description{
Computes the total phi and total theta vector length.
}
