% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scales_cpp}
\alias{scales_cpp}
\title{Computes the MODWT scales}
\usage{
scales_cpp(nb_level)
}
\arguments{
\item{nb_level}{A \code{integer} that contains the level of decomposition J.}
}
\value{
A \code{vec} that contains 2^1, ... , 2^J
}
\description{
Calculates the MODWT scales
}
\details{
Used in wvar object.
}
\examples{
scales_cpp(5)
}
\keyword{internal}
