% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection_methods.R
\name{select.desc.check}
\alias{select.desc.check}
\title{TS Model Checks}
\usage{
select.desc.check(desc)
}
\arguments{
\item{desc}{A \code{character} vector containing \code{ts.model} terms from desc.}
}
\value{
Returns nothing if there is no issues. Otherwise, it will kill the process.
}
\description{
Stops the process in R if there is an issue with the model desc
}
\details{
Checks if there are two or more objects found of type: DR, QN, RW, or WN. In addition, it currently forbids ARMA Models.
}
\keyword{internal}

