% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{select_filter}
\alias{select_filter}
\title{Select the Wavelet Filter}
\usage{
select_filter(filter_name)
}
\arguments{
\item{filter_name}{A \code{String} that must receive: \code{"haar"}.}
}
\value{
info A \code{field<vec>} that contains:
\itemize{
 \item{"L"}{A \code{integer} specifying the length of the filter}
 \item{"h"}{A \code{vector} containing the coefficients for the wavelet filter}
 \item{"g"}{A \code{vector} containing the coefficients for the scaling filter}
}
}
\description{
Constructs the wavelet filter to be used.
}
\details{
The package is oriented toward using only the haar filter. If the package extends at a later time, then the supporting infrastructure is there.
}
\examples{
select_filter("haar")
}
\author{
JJB
}
\keyword{internal}
