% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{seq_cpp}
\alias{seq_cpp}
\title{Generate a sequence of values}
\usage{
seq_cpp(a, b)
}
\arguments{
\item{a}{An \code{int}, that denotes the starting point.}

\item{b}{An \code{int}, that denotes the ending point.}
}
\value{
A \code{vector} containing values moving from a to b. There are no restrictions on A's range.
}
\description{
Creates a vector containing a sequence of values starting at the initial point and going to the terminal point.
}
\examples{
#Call with the following data:
seq_cpp(3, 5)
seq_cpp(5, 3)
}
\author{
James J Balamuta
}
\keyword{internal}
