% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{seq_len_cpp}
\alias{seq_len_cpp}
\title{Generate a sequence of values based on supplied number}
\usage{
seq_len_cpp(n)
}
\arguments{
\item{n}{An \code{int} that denotes the length of the vector.}
}
\value{
A \code{vector} containing values moving from 1 to n.
}
\description{
Creates a vector containing a sequence of values starting at 1 and going to the terminal point.
}
\examples{
#Call with the following data:
seq_len_cpp(5)
}
\author{
James J Balamuta
}
\keyword{internal}
