% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sort_mat}
\alias{sort_mat}
\title{Sort Matrix by Column}
\usage{
sort_mat(x, col)
}
\arguments{
\item{x}{A \code{matrix} to sort}

\item{col}{A \code{int} that indicates the column the matrix should sort by.}
}
\value{
The matrix sorted by values in the specified column.
}
\description{
Sorts a given matrix by a specific column while retain the elements in each row.
}
\details{
The functional difference between armadillo's sort() and sort_mat() is straight forward.
sort() will sort each column without respect to the rows. 
Using sort_matrix will sort only 1 column and retain the other elements to be in the same row.
}
\keyword{internal}
