% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection_methods.R
\name{summary.auto.imu}
\alias{summary.auto.imu}
\title{Summary function for auto.imu object}
\usage{
\method{summary}{auto.imu}(object, digits = 4, ...)
}
\arguments{
\item{object}{A \code{auto.imu} object.}

\item{digits}{A \code{int} indicating how the numbers should be rounded.}

\item{...}{Additional parameters}
}
\description{
Provides a summary of the auto.imu results.
}

