% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avar.R
\name{summary.avar}
\alias{summary.avar}
\title{Summary Allan Variance}
\usage{
\method{summary}{avar}(object, ...)
}
\arguments{
\item{object}{A \code{avar} object.}

\item{...}{Additional arguments affecting the summary produced.}
}
\value{
Summary table
}
\description{
Displays the summary table of allan variance
}
\examples{
# Set seed for reproducibility
set.seed(999)

# Generate time series
x = gen_gts(100, WN(sigma2 = 1))

# Compute Allan
out = avar(x)

# Summary
summary( out )
}
\author{
JJB
}

