% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GMWM.R
\name{summary.gmwm}
\alias{summary.gmwm}
\title{Summary of GMWM object}
\usage{
\method{summary}{gmwm}(object, ci.bootstrap = NULL, ci.b = 20,
  optim.bootstrap = NULL, o.b = 100, ...)
}
\arguments{
\item{object}{A \code{GMWM} object}

\item{ci.bootstrap}{A value containing either: NULL (auto), TRUE, or FALSE}

\item{...}{other arguments passed to specific methods}

\item{model.select}{A value containing either: NULL (auto), TRUE, FALSE}
}
\value{
Text output via print
}
\description{
Displays summary information about GMWM object
}
\examples{
# AR
set.seed(1336)
n = 200
xt = gen.ts(AR1(phi=.1, sigma2 = 1) + AR2(phi=0.95, sigma2 = .1),n)
mod = gmwm(AR1()+AR1(), data=xt, model.type="imu")
summary(mod)
}
\author{
JJB
}

