% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modwt.R
\name{summary.gmwm_modwt}
\alias{summary.gmwm_modwt}
\title{Summary Maximum Overlap Discrete Wavelet Transform}
\usage{
\method{summary}{gmwm_modwt}(object, ...)
}
\arguments{
\item{object}{A \code{gmwm_modwt} object}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
Prints the modwt matrix decomposition
}
\description{
Unlists MODWT object and places it in matrix form
}
\examples{
set.seed(999)
x=rnorm(100)
summary(modwt(x))
}
\author{
JJB
}
\keyword{internal}

