% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadam.R
\name{summary.hadam}
\alias{summary.hadam}
\title{Summary Hadamard Variance}
\usage{
\method{summary}{hadam}(object, ...)
}
\arguments{
\item{object}{A \code{hadam} object.}

\item{...}{Additional arguments affecting the summary produced.}
}
\value{
Summary table
}
\description{
Displays the summary table of Hadamard variance
}
\examples{
set.seed(999)
x = rnorm(100)
out = hadam(x)
summary( out )
}
\author{
JJB
}

