% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modwt.R
\name{summary.modwt}
\alias{summary.modwt}
\title{Summary Maximum Overlap Discrete Wavelet Transform}
\usage{
summary.modwt(x, ...)
}
\arguments{
\item{...}{Not used}

\item{x}{A \code{modwt} object}
}
\value{
Prints the modwt matrix decomposition
}
\description{
Unlists MODWT object and places it in matrix form
}
\examples{
set.seed(999)
x=rnorm(100)
summary(modwt(x))
}
\author{
JJB
}

