% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection_methods.R
\name{summary.rank.models}
\alias{summary.rank.models}
\title{Summary function for rank.models object}
\usage{
\method{summary}{rank.models}(object, digits = 4, ...)
}
\arguments{
\item{object}{A \code{rank.models} object.}

\item{digits}{A \code{int} indicating how the numbers should be rounded.}

\item{...}{Additional parameters}
}
\description{
Provides a summary of the rank.models results.
}

