% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wvar.R
\name{summary.wvar.imu}
\alias{summary.wvar.imu}
\title{Summary of Wavelet Variances for \code{imu} Object}
\usage{
\method{summary}{wvar.imu}(object, ...)
}
\arguments{
\item{object}{A \code{wvar.imu} object.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
Summary table and other properties of the object.
}
\description{
Displays the summary table of wavelet variance in addition to CI values and
supplied efficiency for \code{imu} Object.
}
\examples{
\dontrun{
if(!require("imudata")){
install_imudata()
library("imudata")
}

data(imu6)

test = imu(imu6, gyros = 1:3, accels = 4:6, axis = c('x', 'y', 'z'), freq = 100)
x = wvar(test)
summary(x)
}
}
\keyword{internal}

