% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wvar.R
\name{summary.wvar}
\alias{summary.wvar}
\title{Summary of Wavelet Variances}
\usage{
summary.wvar(object, ...)
}
\arguments{
\item{object}{A \code{wvar} object.}

\item{...}{Not used...}
}
\value{
Summary table and other properties of the object.
}
\description{
Displays the summary table of wavelet variance in addition to CI values and supplied efficiency.
}
\examples{
set.seed(999)
x=rnorm(100)
out = wvar(modwt(x))
summary( out )
}
\author{
JJB
}

