% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compute.cov.R
\name{summary.wvcov}
\alias{summary.wvcov}
\title{Summary Maximum Overlap Discrete Wavelet Transform}
\usage{
summary.modwt(object, ...)
}
\arguments{
\item{object}{A \code{wvcov} object}

\item{...}{Not used}
}
\value{
Prints the modwt matrix decomposition
}
\description{
Unlists MODWT object and places it in matrix form
}
\examples{
set.seed(999)
x=rnorm(100)
decomp = modwt(x)
wv = wvar(decomp)
summary(wvcov(decomp,wv,compute.v="diag"))
}
\author{
JJB
}

