% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{theta_ci}
\alias{theta_ci}
\title{Generate the Confidence Interval for Theta Estimates}
\usage{
theta_ci(theta, psi, alpha)
}
\arguments{
\item{theta}{}

\item{psi}{}

\item{alpha}{}
}
\value{
A \code{mat} that has the first column
}
\description{
yaya
}

