% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.model.R
\name{*.ts.model}
\alias{*.ts.model}
\title{Multiple a ts.model by constant}
\usage{
\method{*}{ts.model}(x, y)
}
\arguments{
\item{x}{A \code{numeric} value}

\item{y}{A \code{ts.model} object}
}
\value{
An S3 object with called ts.model with the following structure:
\itemize{
 \item{process.desc}{y desc replicated x times}
 \item{theta}{y theta replicated x times}
 \item{plength}{Number of Parameters}
 \item{desc}{y desc replicated x times}
 \item{obj.desc}{Depth of Parameters e.g. list(c(1,1),c(1,1))}
 \item{starting}{Guess Starting values? TRUE or FALSE (e.g. specified value)}
}
}
\description{
Sets up the necessary backend for creating multiple model objects.
}
\examples{
4*DR()+2*WN()
DR()*4 + WN()*2
AR1(phi=.3,sigma=.2)*3
}
\author{
JJB
}
\keyword{internal}

