% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{transform_values}
\alias{transform_values}
\title{Transform Values for Optimization}
\usage{
transform_values(theta, desc, objdesc, model_type)
}
\value{
A \code{vec} containing the transformed guesses.
}
\description{
Transform parameter guesses prior to estimating with GMWM
}

