% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{untransform_values}
\alias{untransform_values}
\title{Revert Transform Values for Display}
\usage{
untransform_values(theta, desc, objdesc, model_type)
}
\value{
A \code{vec} containing the undone transformation of parameters.
}
\description{
Undo the previous transform of parameter guesses to obtain the GMWM estimates.
}

