% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_obj}
\alias{update_obj}
\title{Update the Attributes of Objects}
\usage{
update_obj(object, type, new, keep.start = T)
}
\arguments{
\item{object}{A \code{lts}, \code{gts} or \code{imu} object}

\item{type}{A \code{string} that contains the attribute to be updated}

\item{new}{The updated value for the attribute}

\item{keep.start}{A \code{boolean} value that indicates whether 'start' or 'end' should remain the same when 'freq' is updated}
}
\value{
An object with the updated attribute.
}
\description{
Internal Function to Update the Attributes of Objects
}
\keyword{internal}
