% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{value}
\alias{value}
\alias{value.imu}
\title{Obtain the value of an object's properties}
\usage{
value(x, type)

\method{value}{imu}(x, type)
}
\arguments{
\item{x}{A \code{gts}, \code{imu}, or \code{lts} object.}

\item{type}{A \code{string} indicating the field to be retrieved.}
}
\value{
The method will return a single numeric or string result depending on the
slot being accessed.
}
\description{
Used to access different properties of the
 \code{gts}, \code{imu}, or \code{lts} object.
}
\details{
To access information about \code{imu} properties use:
\describe{
 \item{\code{"accel"}}{Returns the number of accelerometers}
 \item{\code{"gyro"}}{Returns the number of gyroscopes}
 \item{\code{"sensors"}}{Returns total number of sensors}
}
}
\section{Methods (by class)}{
\itemize{
\item \code{imu}: Access \code{imu} object properties
}}
\author{
JJB
}

