% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wave_variance}
\alias{wave_variance}
\title{Generate a Confidence intervval for a Univariate Time Series}
\usage{
wave_variance(x, type = "eta3", p = 0.025)
}
\arguments{
\item{x}{A \code{field<vec>} that contains the brick walled modwt or dwt decomposition}

\item{type}{A \code{String} indicating the confidence interval being calculated.}

\item{p}{A \code{double} that indicates the \eqn{\left(1-p\right)*\alpha}{(1-p)*alpha} confidence level}
}
\value{
A \code{matrix} with the structure:
\itemize{
 \item{Column 1}{Wavelet Variance}
 \item{Column 2}{Chi-squared Lower Bounds}
 \item{Column 3}{Chi-squared Upper Bounds}
}
}
\description{
Computes an estimate of the multiscale variance and a chi-squared confidence interval
}
\details{
This function can be expanded to allow for other confidence interval calculations.
}
\examples{
x=rnorm(100)
wave_variance(brick_wall(modwt_cpp(x), haar_filter()))
}

