% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webui.R
\name{webui}
\alias{webui}
\title{Launch Web Modeling UI}
\usage{
webui()
}
\description{
Creates a Web UI Modeling Environment to work with the GMWM objects
}
\examples{
\dontrun{
# Launch the modeling UI
webui()
}
}

