% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ts.model.R
\name{WN}
\alias{WN}
\title{Create an White Noise (WN) Process}
\usage{
WN(sigma2 = NULL)
}
\arguments{
\item{sigma2}{A \code{double} value for the variance, \eqn{\sigma ^2}{sigma^2}, of a WN process.}
}
\value{
An S3 object with called ts.model with the following structure:
\itemize{
 \item{desc}{"WN"}
 \item{theta}{\eqn{\sigma}{sigma}}
 \item{plength}{Number of Parameters}
 \item{obj}{Depth of Parameters e.g. list(1)}
 \item{adv}{Advanced Input TRUE or FALSE (e.g. specified value)}
}
}
\description{
Sets up the necessary backend for the WN process.
}
\examples{
WN()
WN(sigma=3.4)
}
\author{
JJB
}

