% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/analytical_matrix_derivatives.cpp,
%   src/analytical_matrix_derivatives.h
\name{arma_adapter}
\alias{arma_adapter}
\title{ARMA Adapter to ARMA to WV Process function}
\usage{
arma_adapter(theta, p, q, tau)
}
\arguments{
\item{theta}{A \code{vec} that contains all the parameter estimates.}

\item{p}{A \code{int} that indicates the number of AR coefficients}

\item{q}{A \code{int} that indicates the number of MA coefficients.}

\item{tau}{A \code{vec} containing the scales e.g. \eqn{2^{\tau}}{2^tau}}
}
\value{
A \code{vec} containing the ARMA to WV results
}
\description{
Molds the data so that it works with the arma_to_wv function.
}
\details{
The data conversion is more or less a rearrangement of values without using the obj desc.
}
\author{
James Joseph Balamuta (JJB)
}
\keyword{internal}
