% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/process_to_wv.cpp, src/process_to_wv.h
\name{arma_to_wv}
\alias{arma_to_wv}
\title{ARMA process to WV}
\usage{
arma_to_wv(ar, ma, sigma2, tau)
}
\arguments{
\item{ar}{A \code{vec} containing the coefficients of the AR process}

\item{ma}{A \code{vec} containing the coefficients of the MA process}

\item{sigma2}{A \code{double} containing the residual variance}

\item{tau}{A \code{vec} containing the scales e.g. \eqn{2^{\tau}}{2^tau}}
}
\value{
A \code{vec} containing the wavelet variance of the ARMA process.
}
\description{
This function computes the Haar Wavelet Variance of an ARMA process
}
\details{
The function is a generic implementation that requires a stationary theoretical autocorrelation function (ACF)
and the ability to transform an ARMA(\eqn{p},\eqn{q}) process into an MA(\eqn{\infty}{infinity}) (e.g. infinite MA process).
}
\section{Process Haar Wavelet Variance Formula}{

The Autoregressive Order \eqn{p} and Moving Average Order \eqn{q} (ARMA(\eqn{p},\eqn{q})) process has a Haar Wavelet Variance given by:
\deqn{\frac{{{\tau _j}\left[ {1 - \rho \left( {\frac{{{\tau _j}}}{2}} \right)} \right] + 2\sum\limits_{i = 1}^{\frac{{{\tau _j}}}{2} - 1} {i\left[ {2\rho \left( {\frac{{{\tau _j}}}{2} - i} \right) - \rho \left( i \right) - \rho \left( {{\tau _j} - i} \right)} \right]} }}{{\tau _j^2}}\sigma _X^2}{(tau[j]*(1-rho(tau[j]/2)) + 2*sum(i*(2*rho(tau[j]/2 - i) + rho(i) - rho(tau[j] - i))))/tau[j]^2 * sigma[x]^2}
 where \eqn{\sigma _X^2}{sigma[X]^2} is given by the variance of the ARMA process. 
Furthermore, this assumes that stationarity has been achieved as it directly
}

\section{Haar Wavelet Derivation Information}{

For more information, please see: \href{http://smac-group.com/computing/process-to-haar-wavelet-variance-formulae}{Supported Haar Wavelet Formulae} (Internet Connection Required).
}

\examples{
# Calculates the Haar WV for an ARMA(2,3).
wv.theo = arma_to_wv(c(.23,.43), c(.34,.41,.59), 3, 2^(1:9))
}
\seealso{
\code{\link{ARMAtoMA_cpp}}, \code{\link{ARMAacf_cpp}}, and \code{\link{arma11_to_wv}}
}
