% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GMWM.R
\name{autoplot.gmwm2}
\alias{autoplot.gmwm2}
\title{Graph Solution of the Generalized Method of Wavelet Moments}
\usage{
\method{autoplot}{gmwm2}(object, CI = T, transparence = 0.1,
  color.line = c("#003C7D", "#003C7D"), color.CI = "#003C7D",
  line.type = c("solid", "dotted"), graph.title = NA,
  graph.title.size = 15, axis.label.size = 13, axis.tick.size = 11,
  title.x.axis = expression(paste("Scale ", tau)),
  title.y.axis = expression(paste("Wavelet Variance ", nu)),
  legend.title = "", legend.key.size = 1, legend.title.size = 13,
  legend.text.size = 13, ...)
}
\arguments{
\item{object}{A \code{GMWM} object.}

\item{CI}{A \code{boolean} that indicates whether the confidence interval should be plotted}

\item{transparence}{A \code{double} that ranges from 0 to 1 that controls the transparency of the graph}

\item{color.line}{A \code{vector} of \code{string} that indicates the color of the lines for Empirical WV and CI. Default value is \code{c("#003C7D", "#003C7D")}}

\item{color.CI}{A \code{string} that indicates the color of the confidence interval (e.g. black, red, #003C7D, etc.)}

\item{line.type}{A \code{vector} that indicates the type of lines for Empirical WV and its CI. Default value is \code{c("solid","dotted")}}

\item{graph.title}{A \code{string} that indicates the title of the graph}

\item{graph.title.size}{An \code{integer} that indicates the size of title.}

\item{axis.label.size}{An \code{integer} that indicates the size of label}

\item{axis.tick.size}{An \code{integer} that indicates the size of tick mark}

\item{title.x.axis}{A \code{string} that indicates the label on x axis}

\item{title.y.axis}{A \code{string} that indicates the label on y axis}

\item{legend.title}{A \code{string} that indicates the title of legend}

\item{legend.key.size}{A \code{double} that indicates the size of key (in centermeters) on legend}

\item{legend.title.size}{An \code{integer} that indicates the size of title on legend}

\item{legend.text.size}{An \code{integer} that indicates the size of key label on legend}

\item{...}{other arguments passed to specific methods}
}
\value{
A ggplot2 panel containing the graph of the empirical and theoretical wavelet variance under the constructed GMWM for each latent process.
}
\description{
Creates a graph containing the empirical and theoretical wavelet variances constructed via GMWM for each latent process.
}
\author{
JJB
}

