% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wvar.R
\name{autoplot.wvarcompSplit}
\alias{autoplot.wvarcompSplit}
\title{Compare Wavelet Variances on Split}
\usage{
autoplot.wvarcomp(object, ...)
}
\arguments{
\item{object}{A \code{data.frame} containing both sets of variances.}

\item{...}{Not used...}
}
\value{
A ggplot2 panel containing two graphs of the wavelet variance.
}
\description{
Creates the wavelet variance graphs with classic and robust
on separate graphs with the same panel
}
\examples{
set.seed(999)
x=rnorm(100)
Classic = wvar(modwt(x))
Robust = wvar(modwt(x), robust=TRUE)
compare.WV(Classic = Classic, Robust = Robust, split = TRUE)
}
\author{
JJB
}

