% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avar.r
\name{avar}
\alias{avar}
\title{Calculate the Allan Variance}
\usage{
avar(x)
}
\arguments{
\item{x}{A \code{vec} containing the time series under observation.}
}
\value{
Allan variance fixed
}
\description{
Computes the Allan Variance
}
\examples{
set.seed(999)
x=rnorm(100)
avar(x)
}
\author{
JJB
}

