% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/model_selection.cpp, src/model_selection.h
\name{B_matrix}
\alias{B_matrix}
\title{B Matrix}
\usage{
B_matrix(A, at_omega)
}
\arguments{
\item{A}{A \code{mat} containing the first derivatives of the process.}

\item{at_omega}{A \code{mat} containing A^T * Omega}
}
\value{
A \code{mat}
}
\description{
B Matrix
}
\keyword{internal}
