% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{boot_pval_gof}
\alias{boot_pval_gof}
\title{Generate the Confidence Interval for GOF Bootstrapped}
\usage{
boot_pval_gof(obj, obj_boot, B = 1000L, alpha = 0.05)
}
\arguments{
\item{obj}{A \code{double} containing the objective value of the solution.}

\item{obj_boot}{A \code{vec} containing the objective values obtained while bootstrapping under Theta_hat.}

\item{B}{An \code{int} indicating how many times the bootstrapper should be run.}

\item{alpha}{A \code{double} indicating the amount of confidence for CI}
}
\value{
A \code{vec} that has the alpha/2.0 quantile and then the 1-alpha/2.0 quantile.
}
\description{
yaya
}
\keyword{internal}
