% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculate_psi_matrix}
\alias{calculate_psi_matrix}
\title{Calculate the Psi matrix}
\usage{
calculate_psi_matrix(A, v_hat, omega)
}
\arguments{
\item{A}{first derivative matrix}

\item{v_hat}{bootstrapped V}

\item{omega}{original omega matrix}
}
\value{
A \code{mat} that has the first column
}
\description{
Computes the Psi matrix using supplied parameters
}

