% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gmwm.R
\name{compare.models}
\alias{compare.models}
\title{Graphically Compare GMWM Model Fit Between Two Models}
\usage{
compare.models(GMWM1, GMWM2, split = FALSE)
}
\arguments{
\item{GMWM1}{A \code{gmwm} object}

\item{GMWM2}{A \code{gmwm} object}

\item{split}{A \code{boolean} indicating true or false to place model fits on different or the same graphs.}
}
\value{
A ggplot2 panel containing two graphs of the wavelet variance.
}
\description{
Creates GMWM model fits of two models on split graphs within the same panel.
}
\examples{
# AR
set.seed(1336)
n = 200
x = gen_ar1(n, phi=.1, sig2 = 1) + gen_ar1(n,phi=0.95, sig2 = .1)
decomp = modwt(x)
wv = wvar(decomp, robust = TRUE)
out = wvcov(decomp, wv, compute.v="diag")
GMWM1 = gmwm(AR1(), wvcov=out, signal=x, model.type="imu", B = 10000)
GMWM2 = gmwm(2*AR1(), wvcov=out, signal=x, model.type="imu", B = 10000)
compare.models(GMWM1, GMWM2, split = FALSE)
}
\author{
JJB
}

