% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GMWM.R
\name{compare.models}
\alias{compare.models}
\title{Graphically Compare GMWM Model Fit Between Two Models}
\usage{
compare.models(GMWM1, GMWM2, split = FALSE)
}
\arguments{
\item{GMWM1}{A \code{gmwm} object}

\item{GMWM2}{A \code{gmwm} object}

\item{split}{A \code{boolean} indicating true or false to place model fits on different or the same graphs.}
}
\value{
A ggplot2 panel containing two graphs of the wavelet variance.
}
\description{
Creates GMWM model fits of two models on split graphs within the same panel.
}
\examples{
\dontrun{# AR
set.seed(8836)
n = 200
x = gen.ts(AR1(phi = .1, sigma2 = 1) + AR1(phi = 0.95, sigma2 = .1), n)
GMWM1 = gmwm(AR1(), data = x)
GMWM2 = gmwm(2*AR1(), data = x)
compare.models(GMWM1, GMWM2, split = FALSE)}
}
\author{
JJB
}

