% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wvar.R
\name{compare.wvar}
\alias{compare.wvar}
\title{Compare Wavelet Variances}
\usage{
compare.wvar(Classic, Robust, split = TRUE)
}
\arguments{
\item{Classic}{A \code{wvar} object that has \code{robust=false}.}

\item{Robust}{A \code{wvar} object that has \code{robust=true}.}

\item{split}{A \code{boolean} that indicates whether the graphs should be separate (TRUE)
or graphed ontop of each other (FALSE)}
}
\description{
Compare the estimates given by the classical and robust methods
 of calculating the wavelet variance.
}
\examples{
set.seed(999)
x=rnorm(100)
Classic = wvar(modwt(x))
Robust = wvar(modwt(x), robust=TRUE)
# Graph both the classic and the robust estimator on the same graph
compare.wvar(Classic = Classic, Robust = Robust, split = FALSE)
# Graph both the classic and the robust estimator on split graphs next to each other
compare.wvar(Classic = Classic, Robust = Robust, split = TRUE)
}
\author{
JJB
}

