% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{D_matrix}
\alias{D_matrix}
\title{Analytic D matrix of Processes}
\usage{
D_matrix(theta, desc, objdesc, tau, omegadiff)
}
\arguments{
\item{theta}{A \code{vec} containing the list of estimated parameters.}

\item{desc}{A \code{vector<string>} containing a list of descriptors.}

\item{objdesc}{A \code{field<vec>} containing a list of object descriptors.}

\item{tau}{A \code{vec} containing the scales e.g. \eqn{2^{\tau}}{2^tau}}

\item{omegadiff}{A \code{vec} that contains the result of Omega * (wv_empir - wv_theo)}
}
\value{
A \code{matrix} with the process derivatives going down the column
}
\description{
This function computes each process to WV (haar) in a given model.
}
\details{
Function returns the matrix effectively known as "D"
}
\examples{
# TBA
}
\author{
James Joseph Balamuta (JJB)
}
\keyword{internal}
