% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.model.R
\name{DR}
\alias{DR}
\title{Create an Drift (DR) Process}
\usage{
DR(slope = NULL)
}
\arguments{
\item{slope}{A \code{double} value for the slope of a DR process.}
}
\value{
An S3 object with called ts.model with the following structure:
\itemize{
 \item{process.desc}{Used in summary: "DR"}
 \item{theta}{slope}
 \item{plength}{Number of Parameters}
 \item{obj.desc}{y desc replicated x times}
 \item{obj}{Depth of Parameters e.g. list(1)}
 \item{starting}{Guess Starting values? TRUE or FALSE (e.g. specified value)}
}
}
\description{
Sets up the necessary backend for the DR process.
}
\examples{
DR()
DR(slope=3.4)
}
\author{
JJB
}

