% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gen_ar1}
\alias{gen_ar1}
\title{Generate an AR(1) sequence}
\usage{
gen_ar1(N, phi, sig2)
}
\arguments{
\item{N}{An \code{integer} for signal length.}

\item{phi}{A \code{double} that contains autocorrection.}

\item{sig2}{A \code{double} containing the residual variance.}
}
\value{
gm A \code{vec} containing the AR(1) process.
}
\description{
Generate an AR sequence given phi and sig2.
}
\details{
This needs to be extended to AR(p) see \code{arima.sim} and \code{filter}.
}
\examples{
gen_ar(10, 5, 1.2)
}

