% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gen_rw}
\alias{gen_rw}
\title{Generate a random walk without drift}
\usage{
gen_rw(N, sigma2 = 1)
}
\arguments{
\item{N}{An \code{integer} for signal length.}

\item{sigma2}{A \code{double} that contains process variance.}
}
\value{
grw A \code{vec} containing the random walk without drift.
}
\description{
Generates a random walk without drift.
}
\examples{
gen_rw(10, 8.2)
}

