% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/gen_process.cpp, src/gen_process.h
\name{gen_rw}
\alias{gen_rw}
\title{Generate a random walk without drift}
\usage{
gen_rw(N, sigma2 = 1)
}
\arguments{
\item{N}{An \code{integer} for signal length.}

\item{sigma2}{A \code{double} that contains process variance.}
}
\value{
grw A \code{vec} containing the random walk without drift.
}
\description{
Generates a random walk without drift.
}
\examples{
gen_rw(10, 8.2)
}
\keyword{internal}

