% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/gen_process.cpp, src/gen_process.h
\name{gen_wn}
\alias{gen_wn}
\title{Generate a White Noise Process (\eqn{WN(\sigma^2)})
Generates a White Noise Process with variance parameter \eqn{\sigma ^2}.}
\usage{
gen_wn(N, sigma2 = 1)
}
\arguments{
\item{N}{An \code{integer} for signal length.}

\item{sigma2}{A \code{double} that contains process variance.}
}
\value{
wn A \code{vec} containing the white noise.
}
\description{
Generate a White Noise Process (\eqn{WN(\sigma^2)})
Generates a White Noise Process with variance parameter \eqn{\sigma ^2}.
}
\examples{
gen_wn(10, 1.5)
}
\keyword{internal}

