% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/gen_process.cpp, src/gen_process.h
\name{gen_wn}
\alias{gen_wn}
\title{Generate a white noise process}
\usage{
gen_wn(N, sigma2 = 1)
}
\arguments{
\item{N}{An \code{integer} for signal length.}

\item{sigma2}{A \code{double} that contains process variance.}
}
\value{
wn A \code{vec} containing the white noise.
}
\description{
Generates a white noise process with variance parameter sigma.
}
\examples{
gen_wn(10, 1.5)
}
\keyword{internal}

