% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getEff}
\alias{getEff}
\title{Get \code{gmwm} Efficiency Values}
\usage{
getEff(obj.list)
}
\arguments{
\item{obj.list}{A \code{list} of \code{gmwm} object}
}
\value{
A \code{numeric vector}.
}
\description{
Get efficiency values from a list of \code{gmwm} object
}
\details{
If the object is computed by classical method, it will return 1.1. The reason is:

It's possible for user to create one object by robust method with eff=1, though it is exactly same
as classical method. In this case, if we want the classical method to always appear on the 
top left corner,  number larger than 1 (e.g. 1.1) is used. This setting makes it easy to draw the graph.
}
\examples{
set.seed(8836)
n = 1000
x = gen.gts(AR1(phi = .1, sigma2 = 1) + AR1(phi = 0.95, sigma2 = .1), n)
GMWM1 = gmwm(2*AR1()+RW(), data = x, robust  = FALSE)
GMWM2 = gmwm(2*AR1()+RW(), data = x, robust  = TRUE, eff = 0.1)
GMWM3 = gmwm(2*AR1()+RW(), data = x, robust  = TRUE, eff = 0.6)
obj.list = list(GMWM1, GMWM2, GMWM3)
getEff(obj.list)
}
\keyword{internal}

