% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{ggColor}
\alias{ggColor}
\title{Emulate ggplot2 default color palette}
\usage{
ggColor(n)
}
\arguments{
\item{n}{An \code{integer} indicating how many colors user wants.}
}
\value{
A \code{vector} containing \code{n} colors
}
\description{
Autogenerate a colors according to the ggplot selection mechanism.
}
\author{
John Colby
}
\keyword{internal}

