% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{GMWM-package}
\alias{GMWM-package}
\title{GMWM Package}
\description{
GMWM is a Time Series Package
}
\details{
\tabular{ll}{
Package: \tab GMWM\cr
Type: \tab Package\cr
Version: \tab 0.0.02\cr
Date: \tab 2014-12-05\cr
License: \tab MIT + file\cr
}
Some details should be added
}
\examples{
# Fill in at a later time
}
\author{
Stephane Guerrier \email{stephane@illinois.edu},
James Balamuta \email{balamut2@illinois.edu}

James Balamuta \email{balamut2@illinois.edu}
}
\seealso{
\code{avar}
}

