% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GMWM_package.R
\docType{package}
\name{gmwm-package}
\alias{gmwm-package}
\title{GMWM Package}
\description{
The gmwm package implements the theoretical framework for Generalized Method
of Wavelet Moments (GMWM).
}
\details{
\tabular{ll}{
Package: \tab GMWM\cr
Type: \tab Package\cr
Version: \tab 0.13.0\cr
Date: \tab 2015-11-01\cr
License: \tab file LICENSE\cr
}
}
\author{
James Balamuta \email{balamut2@illinois.edu},
Stephane Guerrier \email{stephane@illinois.edu},
Roberto Molinari \email{roberto.molinari@unige.ch},
Wenchao Yang \email{wyang40@illinois.edu}

Stephane Guerrier \email{stephane@illinois.edu}
}

