% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GMWM_package.R
\docType{package}
\name{GMWM-package}
\alias{GMWM-package}
\title{GMWM Package}
\description{
GMWM is a Time Series Package
}
\details{
\tabular{ll}{
Package: \tab GMWM\cr
Type: \tab Package\cr
Version: \tab 0.12.2\cr
Date: \tab 2015-08-04\cr
License: \tab MIT + file LICENSE\cr
}
Some details should be added
}
\examples{
# Fill in at a later time
}
\author{
James Balamuta \email{balamut2@illinois.edu},
Stephane Guerrier \email{stephane@illinois.edu},
Roberto Molinari \email{roberto.molinari@unige.ch},
Wenchao Yang \email{wyang40@illinois.edu}

Stephane Guerrier \email{stephane@illinois.edu}
}
\seealso{
\code{avar}
}

