% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMWM.R
\name{gmwm.imu}
\alias{gmwm.imu}
\title{GMWM for (Robust) sensor}
\usage{
gmwm.imu(model, data, compute.v = "fast", robust = F, eff = 0.6, ...)
}
\arguments{
\item{model}{A \code{ts.model} object containing one of the allowed models.}

\item{data}{A \code{matrix} or \code{data.frame} object with only column (e.g. \eqn{N \times 1}{ N x 1 }), or a \code{lts} object, or a \code{gts} object.}

\item{compute.v}{A \code{string} indicating the type of covariance matrix solver. "fast", "bootstrap", "asymp.diag", "asymp.comp", "fft"}

\item{robust}{A \code{boolean} indicating whether to use the robust computation (TRUE) or not (FALSE).}

\item{eff}{A \code{double} between 0 and 1 that indicates the efficiency.}

\item{...}{other arguments passed to the main gmwm function}
}
\value{
A \code{gmwm} object that contains:
\itemize{
 \item{}
 \item{}
 \item{}
}
}
\description{
GMM object
}

