% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gmwm.R
\name{gmwm}
\alias{gmwm}
\title{GMWM for IMU, ARMA, SSM, and Robust}
\usage{
gmwm(model, wvcov, signal, model.type="imu", B = 1000)
}
\arguments{
\item{model}{A \code{ts.model} object containing one of the allowed models. Not used if \code{type="arma"}}

\item{wvcov}{A \code{wvcov} object}

\item{signal}{A \code{vec} that is the time series being studied}

\item{model.type}{A \code{string} containing the type of GMWM needed e.g. IMU, ARMA, or SSM}

\item{B}{A \code{integer} to sample the space for IMU and SSM models to ensure AR1 identitability.}
}
\value{
A \code{gmwm} object that contains:
\itemize{
 \item{}
 \item{}
 \item{}
}
}
\description{
GMM object
}
\examples{
# AR
set.seed(1336)
n = 200
x = gen_ar1(n, phi=.1, sigma2 = 1) + gen_ar1(n,phi=0.95, sigma2 = .1)
decomp = modwt(x)
wv = wvar(decomp, robust = FALSE)
out = wvcov(decomp, wv, compute.v="diag")
save1 = gmwm(AR1(), wvcov=out, signal=x, model.type="imu", B = 10000)
save2 = gmwm(2*AR1(), wvcov=out, signal=x, model.type="imu", B = 10000)

# ARMA case
set.seed(1336)
x = arima.sim(n = 200,
              list(ar = c(0.8897, -0.4858), ma = c(-0.2279, 0.2488)),
              sd = sqrt(0.1796))
decomp = modwt(x)
wv = wvar(decomp, robust = TRUE)
out = wvcov(decomp, wv, compute.v="diag")
save = gmwm(ARMA(2,2),wvcov=out, signal=x, model.type="arma")
}

