% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/imu.R
\docType{data}
\name{imu}
\alias{imu}
\title{IMU Data}
\format{A \code{data.frame} with 873,684 observations and 7 variables. The variables are as follows:
\itemize{
 \item{V1}{Time}
 \item{V2}{X-Axis}
 \item{V3}{Y-Axis}
 \item{V4}{Z-Axis}
 \item{V5}{Fill in}
 \item{V6}{Fill in}
 \item{V7}{Fill in}
}}
\source{
Where's the data frame?
}
\usage{
imu
}
\description{
This data set contains gyroscope data
}
\examples{
\dontrun{
data(imu)
}
}
\author{
JJB
}

