% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/imu_graphs.r
\name{imu2WV}
\alias{imu2WV}
\title{Obtain the Wavelet Variance for imu data}
\usage{
imu2WV(object, gyroscope = c(1:3), accelerometer = c(4:6))
}
\arguments{
\item{object}{A \code{data.frame} or \code{matrix} that contains 6 columns.}

\item{gyroscope}{A \code{vector} that contains the index of column where gyroscope data (Gyro. X, Gyro. Y and Gyro. Z) is placed}

\item{accelerometer}{A \code{vector} that contains the index of column where accelerometer data (Accel. X, Accel. Y and Accel. Z) is placed}
}
\value{
An object whose class is \code{imu} that contains the formatted data.
}
\description{
Generates the Wavelet Variance for each column in the data set.
Data set is molten, \code{scales},\code{Axis} and \code{sensor} are used as id variables (see \code{?melt.data.frame})
}
\examples{
\dontrun{
data(imu)
df = imu2WV(imu)
}
}

