% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/analytical_matrix_derivatives.cpp,
%   src/analytical_matrix_derivatives.h
\name{jacobian_arma}
\alias{jacobian_arma}
\title{Calculates the Jacobian for the ARMA process}
\usage{
jacobian_arma(theta, p, q, tau)
}
\arguments{
\item{theta}{A \code{vec} that contains all the parameter estimates.}

\item{p}{A \code{int} that indicates the number of AR coefficients}

\item{q}{A \code{int} that indicates the number of MA coefficients.}

\item{tau}{A \code{vec} containing the scales e.g. \eqn{2^{\tau}}{2^tau}}
}
\value{
A \code{mat} that returns the first numerical derivative of the ARMA process.
}
\description{
Take the numerical derivative for the first derivative of an ARMA using the 2 point rule.
}
\author{
James Joseph Balamuta (JJB)
}
\keyword{internal}
