% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logit}
\alias{logit}
\title{Logit Function}
\usage{
logit(x)
}
\arguments{
\item{x}{A \code{vec} containing probabilities (e.g. -1 <= x <= 1)}
}
\value{
A \code{vec} containing logit terms.
}
\description{
This function computes the logit link function.
}
\examples{
# Set seed for reproducibility
set.seed(1142)

# Simulate data
x.sim = runif(10)

# Transform
x.sim.transformed = logit(x.sim)
}
\seealso{
\code{\link{logit_inv}}
}
\author{
James Joseph Balamuta (JJB)
}
\keyword{internal}
