% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GMWM.R
\name{plot.gmwm}
\alias{plot.gmwm}
\title{Wrapper to Graph Solution of the Generalized Method of Wavelet Moments}
\usage{
\method{plot}{gmwm}(x, ...)
}
\arguments{
\item{x}{A \code{GMWM} object}

\item{...}{other arguments passed to specific methods}
}
\value{
A ggplot2 panel containing the graph of the empirical and theoretical wavelet variance under the constructed GMWM.
}
\description{
Creates a graph containing the empirical and theoretical wavelet variances constructed via GMWM.
}
\examples{
# AR
set.seed(1336)
n = 200
x = gen_ar1(n, phi=.1, sigma2 = 1) + gen_ar1(n,phi=0.95, sigma2 = .1)
mod = gmwm(AR1(), data=x, model.type="imu")
plot(mod)
}
\author{
JJB
}

