% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMWM.R
\name{plot.gmwm}
\alias{plot.gmwm}
\title{plot method for GMWM}
\usage{
\method{plot}{gmwm}(
  x,
  decomp = TRUE,
  add_legend_gwmw = TRUE,
  ylab_gmwm = NULL,
  col_theo = "#F47F24",
  pch_theo = 1,
  cex_theo = 1.25,
  lwd_theo = 1.25,
  add_legend = TRUE,
  legend_pos = NULL,
  col_wv = NULL,
  col_ci = NULL
)
}
\description{
plot method for GMWM
}
