% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/imu_graphs.r
\name{plot.imu}
\alias{plot.imu}
\title{Plot the Wavelet Variances in separate and combined way}
\usage{
\method{plot}{imu}(x, separate = TRUE, transparence = 0.1, point.size = 0,
  color.line = "black", color.point = "#003C7D", color.CI = "#003C7D",
  line.type = "solid", ...)
}
\arguments{
\item{x}{An \code{imu} object}

\item{separate}{A \code{boolean} that indicates whether the graphs should be split or combined.}

\item{transparence}{A \code{double} that ranges from 0 to 1 that controls the transparency of the graph}

\item{point.size}{An \code{integer} that indicates the size of the points.}

\item{color.line}{A \code{string} that indicates the color of the line drawn.}

\item{color.point}{A \code{string} that is a hexadecimal color value.}

\item{color.CI}{A \code{string} that is a hexadecimal color value.}

\item{line.type}{A \code{string} that indicates the type of line (e.g. solid, dotted, etc.)}

\item{...}{Additional options}
}
\value{
A panel containing the graph of an IMU sensor.
}
\description{
Creates a graph of the wavelet variance object. The graphs can be either split (features a CI) or combined (all gyroscopes together and all accelerometer).
}
\examples{
\dontrun{
data(imu)
df = imu2WV(imu)
plot(df)
plot(df, separate=FALSE)
}
}

