% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wvar.R
\name{print.wvar}
\alias{print.wvar}
\title{Print Wavelet Variances}
\usage{
\method{print}{wvar}(x, ...)
}
\arguments{
\item{x}{A \code{wvar} object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Summary table
}
\description{
Displays the summary table of wavelet variance.
}
\examples{
set.seed(999)
x = rnorm(100)
out = wvar(x)
print( out )
}
\author{
JJB
}
\keyword{internal}
