% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wvcov.R
\name{print.wvcov}
\alias{print.wvcov}
\title{Print Asymptotic Covariance Matrix}
\usage{
\method{print}{wvcov}(x, ...)
}
\arguments{
\item{x}{A \code{wvcov} object}

\item{...}{further arguments passed to or from other methods}
}
\value{
Prints the modwt matrix decomposition
}
\description{
Places the Asymptotic Covariance Matrix in print form.
}
\examples{
set.seed(999)
x=rnorm(100)
decomp = modwt(x)
wv = wvar(x)
print(wvcov(decomp,wv,compute.v="diag"))
}
\author{
JJB
}
\keyword{internal}

