% Generated by roxygen2 (4.0.2): do not edit by hand
\name{QN}
\alias{QN}
\title{Create an Quantisation Noise (QN) Process}
\usage{
QN(q2)
}
\arguments{
\item{q2}{A \code{double} value for the \eqn{Q^2}{Q^2} of a QN process.}
}
\value{
An S3 object with called ts.model with the following structure:
\itemize{
 \item{desc}{"QN"}
 \item{theta}{\eqn{Q^2}{Q^2}}
}
}
\description{
Sets up the necessary backend for the QN process.
}
\details{
A standard deviation is required since the model generation statements utilize
randomization functions expecting a standard deviation instead of a variance.
}
\examples{
QN()
QN(q2=3.4)
}
\author{
JJB
}

