% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.model.R
\name{QN}
\alias{QN}
\title{Create an Quantisation Noise (QN) Process}
\usage{
QN(q2 = NULL)
}
\arguments{
\item{q2}{A \code{double} value for the \eqn{Q^2}{Q^2} of a QN process.}
}
\value{
An S3 object with called ts.model with the following structure:
\itemize{
 \item{process.desc}{Used in summary: "QN"}
 \item{theta}{\eqn{Q^2}{Q^2}}
 \item{plength}{Number of Parameters}
 \item{desc}{y desc replicated x times}
 \item{obj.desc}{Depth of Parameters e.g. list(1)}
 \item{starting}{Guess Starting values? TRUE or FALSE (e.g. specified value)}
}
}
\description{
Sets up the necessary backend for the QN process.
}
\examples{
QN()
QN(q2=3.4)
}
\author{
JJB
}

