% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{quantile_cpp}
\alias{quantile_cpp}
\title{Find Quantiles}
\usage{
quantile_cpp(x, probs)
}
\arguments{
\item{x}{A \code{vec} of data}

\item{probs}{A \code{vec} of the quantiles to find.}
}
\value{
A \code{vector} containing the quantiles
}
\description{
Attempts to find quantiles
}
\examples{
#Call with the following data:
quantile_cpp(c(1,2,3,4,5,6,7), c(.25,.5,.75))
quantile(c(1,2,3,4,5,6,7), c(.25,.5,.75))
}
\author{
James J Balamuta
}
\keyword{internal}
