% Generated by roxygen2 (4.0.2): do not edit by hand
\name{RW}
\alias{RW}
\title{Create an Random Walk (RW) Process}
\usage{
RW(sigma=1)
}
\arguments{
\item{sigma}{A \code{double} value for the standard deviation, \eqn{\sigma}{sigma}, of a RW process.}
}
\value{
An S3 object with called ts.model with the following structure:
\itemize{
 \item{desc}{"RW"}
 \item{theta}{\eqn{\sigma}{sigma}}
}
}
\description{
Sets up the necessary backend for the RW process.
}
\details{
A standard deviation is required since the model generation statements utilize
randomization functions expecting a standard deviation instead of a variance.
}
\examples{
RW()
RW(sigma=3.4)
}
\author{
JJB
}

