% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.model.R
\name{RW}
\alias{RW}
\title{Create an Random Walk (RW) Process}
\usage{
RW(sigma2 = NULL)
}
\arguments{
\item{sigma2}{A \code{double} value for the variance, \eqn{\sigma ^2}{sigma^2}, of a WN process.}
}
\value{
An S3 object with called ts.model with the following structure:
\describe{
 \item{process.desc}{Used in summary: "RW"}
 \item{theta}{\eqn{\sigma}{sigma}}
 \item{plength}{Number of Parameters}
 \item{desc}{y desc replicated x times}
 \item{obj.desc}{Depth of Parameters e.g. list(1)}
 \item{starting}{Guess Starting values? TRUE or FALSE (e.g. specified value)}
}
}
\description{
Sets up the necessary backend for the RW process.
}
\examples{
RW()
RW(sigma=3.4)
}
\author{
JJB
}

