% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avar.r
\name{summary.avar}
\alias{summary.avar}
\title{Summary Allan Variance}
\usage{
\method{summary}{avar}(object, ...)
}
\arguments{
\item{object}{A \code{avar} object.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
Summary table
}
\description{
Displays the summary table of allan variance
}
\examples{
set.seed(999)
x=rnorm(100)
out = avar(x)
summary( out )
}
\author{
JJB
}

