% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GMWM.R
\name{summary.gmwm}
\alias{summary.gmwm}
\title{Summary of GMWM object}
\usage{
\method{summary}{gmwm}(object, ...)
}
\arguments{
\item{...}{other arguments passed to specific methods}

\item{x}{A \code{GMWM} object}
}
\value{
Text output via print
}
\description{
Displays summary information about GMWM object
}
\examples{
# AR
set.seed(1336)
n = 200
xt = gen.ts(AR1(phi=.1, sigma2 = 1) + AR2(phi=0.95, sigma2 = .1),n)
mod = gmwm(AR1()+AR1(), data=xt, model.type="imu")
summary(mod)
}
\author{
JJB
}

