% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wvcov.R
\name{summary.wvcov}
\alias{summary.wvcov}
\title{Summary Wavelet Covariance Matrix}
\usage{
\method{summary}{wvcov}(object, ...)
}
\arguments{
\item{object}{A \code{wvcov} object}

\item{...}{Additional arguments affecting the summary produced.}
}
\value{
Prints the modwt matrix decomposition
}
\description{
Prints the Wavelet Covariance Matrix
}
\examples{
set.seed(999)
x = rnorm(100)
decomp = modwt(x)
wv = wvar(x)
summary(wvcov(decomp,wv,compute.v="diag"))
}
\author{
JJB
}
\keyword{internal}

