% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{transform_values}
\alias{transform_values}
\title{Transform Values for Optimization}
\usage{
transform_values(theta, desc, objdesc, model_type)
}
\arguments{
\item{theta}{A \code{vec} with dimensions N x 1 that contains user-supplied initial values for parameters}

\item{desc}{A \code{vector<string>} indicating the models that should be considered.}

\item{objdesc}{A \code{field<vec>} containing a list of parameters (e.g. AR(1) = c(1,1), ARMA(p,q) = c(p,q,1))}

\item{model_type}{A \code{string} that contains the model type: \code{"imu"} or \code{"ssm"}}
}
\value{
A \code{vec} containing the transformed guesses.
}
\description{
Transform parameter guesses prior to estimating with GMWM
}
\author{
James Joseph Balamuta (JJB)
}
\keyword{internal}
