% Generated by roxygen2 (4.0.2): do not edit by hand
\name{WN}
\alias{WN}
\title{Create an White Noise (WN) Process}
\usage{
WN(sigma=1)
}
\arguments{
\item{sigma}{A \code{double} value for the standard deviation, \eqn{\sigma}{sigma}, of a WN process.}
}
\value{
An S3 object with called ts.model with the following structure:
\itemize{
 \item{desc}{"WN"}
 \item{theta}{\eqn{\sigma}{sigma}}
}
}
\description{
Sets up the necessary backend for the WN process.
}
\details{
A standard deviation is required since the model generation statements utilize
randomization functions expecting a standard deviation instead of a variance.
}
\examples{
WN()
WN(sigma=3.4)
}
\author{
JJB
}

